@@ -11,13 +11,15 @@ import {
   TemplateRef,
   ElementRef,
   AfterContentInit,
-  EventEmitter
+  EventEmitter,
+  HostListener,
+  Inject
 } from '@angular/core';
 
-import { Subscription, Observable, merge } from 'rxjs';
+import { Subscription, Observable, merge, of, from } from 'rxjs';
 
 import { ResizeService } from '../services/resize.service';
-import { tap, delay, filter, switchMap, first } from 'rxjs/operators';
+import { tap, delay, filter, switchMap, first, map, skip, take, toArray } from 'rxjs/operators';
 import { CarouselService, CarouselCurrentData } from '../services/carousel.service';
 import { StageData } from "../models/stage-data.model";
 import { OwlDOMData } from "../models/owlDOM-data.model";
@@ -31,6 +33,7 @@ import { AnimateService } from '../services/animate.service';
 import { AutoHeightService } from '../services/autoheight.service';
 import { HashService } from '../services/hash.service';
 import { OwlLogger } from '../services/logger.service';
+import { DOCUMENT } from '../services/document-ref.service';
 
 let nextId = 0;
 
@@ -139,6 +142,7 @@ export class CarouselComponent
   @Output() translated = new EventEmitter<SlidesOutputData>();
   @Output() dragging = new EventEmitter<{dragging: boolean, data: SlidesOutputData}>();
   @Output() change = new EventEmitter<SlidesOutputData>();
+  @Output() changed = new EventEmitter<SlidesOutputData>();
   @Output() initialized = new EventEmitter<SlidesOutputData>();
 
   /**
@@ -224,6 +228,11 @@ export class CarouselComponent
   private _changeCarousel$: Observable<string>;
 
   /**
+   * Observable for catching the moment when the data about slides changed, more exactly when the position changed.
+   */
+  private _changedCarousel$: Observable<any>;
+
+  /**
    * Observable for catching the initialization of changing the carousel
    */
   private _initializedCarousel$: Observable<string>;
@@ -232,6 +241,7 @@ export class CarouselComponent
    * Observable for merging all Observables and creating one subscription
    */
   private _carouselMerge$: Observable<CarouselCurrentData | string>;
+  private docRef: Document;
 
   constructor(
     private el: ElementRef,
@@ -243,8 +253,30 @@ export class CarouselComponent
     private animateService: AnimateService,
     private autoHeightService: AutoHeightService,
     private hashService: HashService,
-    private logger: OwlLogger
-  ) {}
+    private logger: OwlLogger,
+    @Inject(DOCUMENT) docRef: any
+  ) {
+    this.docRef = docRef as Document;
+
+  }
+
+  @HostListener('document:visibilitychange', ['$event'])
+  onVisibilityChange(ev: Event) {
+    if (!this.carouselService.settings.autoplay) return;
+    switch (this.docRef.visibilityState) {
+      case 'visible':
+        this.autoplayService.play();
+        break;
+
+      case 'hidden':
+        this.autoplayService.stop();
+        break;
+
+      default:
+        break;
+    }
+  };
+
 
   ngOnInit() {
     this.spyDataStreams();
@@ -335,22 +367,80 @@ export class CarouselComponent
       })
     );
 
+    this._changedCarousel$ = this.carouselService.getChangeState().pipe(
+      switchMap(value => {
+        const changedPosition: Observable<SlidesOutputData> = of(value).pipe(
+          filter(() => value.property.name === 'position'),
+          switchMap(() => from(this.slidesData)),
+          skip(value.property.value),
+          take(this.carouselService.settings.items),
+          map(slide => {
+            const clonedIdPrefix = this.carouselService.clonedIdPrefix;
+            const id = slide.id.indexOf(clonedIdPrefix) >= 0 ? slide.id.slice(clonedIdPrefix.length) : slide.id;
+            return { ...slide, id: id, isActive: true };
+          }),
+          toArray(),
+          map(slides => {
+            return {
+              slides: slides,
+              startPosition: this.carouselService.relative(value.property.value)
+            }
+          })
+        );
+
+        // const changedSetting: Observable<SlidesOutputData> = of(value).pipe(
+        //   filter(() => value.property.name === 'settings'),
+        //   map(() => {
+        //     return {
+        //       slides: [],
+        //       startPosition: this.carouselService.relative(value.property.value)
+        //     }
+        //   })
+        // )
+        return merge(changedPosition);
+      }),
+      tap(slidesData => {
+        this.gatherTranslatedData();
+        this.changed.emit(slidesData.slides.length ? slidesData : this.slidesOutputData);
+        // console.log(this.slidesOutputData);
+        // this.slidesOutputData = {};
+      })
+    );
+
     this._draggingCarousel$ = this.carouselService.getDragState().pipe(
       tap(() => {
         this.gatherTranslatedData();
         this.dragging.emit({dragging: true, data: this.slidesOutputData});
       }),
       switchMap(
-        () => this.carouselService.getTranslatedState().pipe(
-          first(),
-          tap(() => {
-            this.dragging.emit({dragging: false, data: this.slidesOutputData});
-          })
+        () => this.carouselService.getDraggedState().pipe(
+          map(() => !!this.carouselService.is('animating'))
         )
-      )
+      ),
+      switchMap(
+        anim => {
+          if (anim) {
+            return this.carouselService.getTranslatedState().pipe(
+              first(),
+            );
+          } else {
+            return of('not animating');
+          }
+        }
+      ),
+      tap(() => {
+        this.dragging.emit({dragging: false, data: this.slidesOutputData});
+      })
     );
 
-    this._carouselMerge$ = merge(this._viewCurSettings$, this._translatedCarousel$, this._draggingCarousel$, this._changeCarousel$, this._initializedCarousel$);
+    this._carouselMerge$ = merge(
+      this._viewCurSettings$,
+      this._translatedCarousel$,
+      this._draggingCarousel$,
+      this._changeCarousel$,
+      this._changedCarousel$,
+      this._initializedCarousel$
+    );
     this._allObservSubscription = this._carouselMerge$.subscribe(() => {});
   }
 
@@ -382,7 +472,7 @@ export class CarouselComponent
    * Handler for click event, attached to next button
    */
   next() {
-    if (!this.carouselLoaded || (this.navData && this.navData.disabled)) return;
+    if (!this.carouselLoaded) return;
     this.navigationService.next(this.carouselService.settings.navSpeed);
   }
 
@@ -390,7 +480,7 @@ export class CarouselComponent
    * Handler for click event, attached to prev button
    */
   prev() {
-    if (!this.carouselLoaded || (this.navData && this.navData.disabled)) return;
+    if (!this.carouselLoaded) return;
     this.navigationService.prev(this.carouselService.settings.navSpeed);
   }
 
@@ -407,7 +497,8 @@ export class CarouselComponent
    * @param id fragment of url
    */
   to(id: string) {
-    if (!this.carouselLoaded || (this.navData && this.navData.disabled) || (this.dotsData && this.dotsData.disabled)) return;
+    // if (!this.carouselLoaded || ((this.navData && this.navData.disabled) && (this.dotsData && this.dotsData.disabled))) return;
+    if (!this.carouselLoaded) return;
     this.navigationService.toSlideById(id);
   }
 
