@@ -1,14 +1,14 @@
 import { EventManager } from '@angular/platform-browser';
-import { Subject, merge } from 'rxjs';
-import { tap, filter, skip, delay, switchMap, first } from 'rxjs/operators';
+import { Subject, merge, of, from } from 'rxjs';
+import { tap, filter, switchMap, first, skip, take, delay, map, toArray } from 'rxjs/operators';
 import { trigger, state, style, animate, transition } from '@angular/animations';
 import { isPlatformBrowser, LocationStrategy, CommonModule } from '@angular/common';
-import { Injectable, ErrorHandler, isDevMode, InjectionToken, PLATFORM_ID, Inject, Component, Input, Output, Directive, ContentChildren, TemplateRef, ElementRef, EventEmitter, NgZone, HostListener, Renderer2, Attribute, HostBinding, NgModule } from '@angular/core';
-import { ActivatedRoute, Router, NavigationEnd, RouterModule } from '@angular/router';
+import { Injectable, ErrorHandler, isDevMode, InjectionToken, PLATFORM_ID, Inject, Optional, Component, Input, Output, Directive, ContentChildren, TemplateRef, ElementRef, EventEmitter, HostListener, NgZone, Renderer2, Attribute, HostBinding, NgModule } from '@angular/core';
+import { ActivatedRoute, Router, NavigationEnd } from '@angular/router';
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 class ResizeService {
     /**
@@ -29,7 +29,6 @@ class ResizeService {
     }
     /**
      * Handler of 'resize' event. Passes data throw resizeSubject
-     * @private
      * @param {?} event Event Object of 'resize' event
      * @return {?}
      */
@@ -38,7 +37,6 @@ class ResizeService {
     }
     /**
      * Handler of 'onload' event. Defines the width of window
-     * @private
      * @param {?} event Event Object of 'onload' event
      * @return {?}
      */
@@ -49,13 +47,14 @@ class ResizeService {
 ResizeService.decorators = [
     { type: Injectable }
 ];
+/** @nocollapse */
 ResizeService.ctorParameters = () => [
     { type: EventManager }
 ];
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 /**
  * Defaults value of options
@@ -87,6 +86,7 @@ class OwlCarouselOConfig {
         this.navText = ['prev', 'next'];
         this.navSpeed = false;
         this.slideBy = 1; // stage moves on 1 width of slide; if slideBy = 2, stage moves on 2 widths of slide
+        // stage moves on 1 width of slide; if slideBy = 2, stage moves on 2 widths of slide
         this.dots = true;
         this.dotsEach = false;
         this.dotsData = false;
@@ -140,6 +140,7 @@ class OwlOptionsMockedTypes {
         this.navText = 'string[]';
         this.navSpeed = 'number|boolean';
         this.slideBy = 'number|string'; // stage moves on 1 width of slide; if slideBy = 2, stage moves on 2 widths of slide
+        // stage moves on 1 width of slide; if slideBy = 2, stage moves on 2 widths of slide
         this.dots = 'boolean';
         this.dotsEach = 'number|boolean';
         this.dotsData = 'boolean';
@@ -164,7 +165,7 @@ class OwlOptionsMockedTypes {
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 class OwlLogger {
     /**
@@ -202,13 +203,14 @@ class OwlLogger {
 OwlLogger.decorators = [
     { type: Injectable }
 ];
+/** @nocollapse */
 OwlLogger.ctorParameters = () => [
     { type: ErrorHandler }
 ];
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 /** @enum {string} */
 const Type = {
@@ -228,52 +230,52 @@ class CarouselService {
     constructor(logger) {
         this.logger = logger;
         /**
-       * Subject for passing data needed for managing View
-       */
+         * Subject for passing data needed for managing View
+         */
         this._viewSettingsShipper$ = new Subject();
         /**
-       * Subject for notification when the carousel got initializes
-       */
+         * Subject for notification when the carousel got initializes
+         */
         this._initializedCarousel$ = new Subject();
         /**
-       * Subject for notification when the carousel's settings start changinf
-       */
+         * Subject for notification when the carousel's settings start changinf
+         */
         this._changeSettingsCarousel$ = new Subject();
         /**
-       * Subject for notification when the carousel's settings have changed
-       */
+         * Subject for notification when the carousel's settings have changed
+         */
         this._changedSettingsCarousel$ = new Subject();
         /**
-       * Subject for notification when the carousel starts translating or moving
-       */
+         * Subject for notification when the carousel starts translating or moving
+         */
         this._translateCarousel$ = new Subject();
         /**
-       * Subject for notification when the carousel stopped translating or moving
-       */
+         * Subject for notification when the carousel stopped translating or moving
+         */
         this._translatedCarousel$ = new Subject();
         /**
-       * Subject for notification when the carousel's rebuilding caused by 'resize' event starts
-       */
+         * Subject for notification when the carousel's rebuilding caused by 'resize' event starts
+         */
         this._resizeCarousel$ = new Subject();
         /**
-       * Subject for notification  when the carousel's rebuilding caused by 'resize' event is ended
-       */
+         * Subject for notification  when the carousel's rebuilding caused by 'resize' event is ended
+         */
         this._resizedCarousel$ = new Subject();
         /**
-       * Subject for notification when the refresh of carousel starts
-       */
+         * Subject for notification when the refresh of carousel starts
+         */
         this._refreshCarousel$ = new Subject();
         /**
-       * Subject for notification when the refresh of carousel is ended
-       */
+         * Subject for notification when the refresh of carousel is ended
+         */
         this._refreshedCarousel$ = new Subject();
         /**
-       * Subject for notification when the dragging of carousel starts
-       */
+         * Subject for notification when the dragging of carousel starts
+         */
         this._dragCarousel$ = new Subject();
         /**
-       * Subject for notification when the dragging of carousel is ended
-       */
+         * Subject for notification when the dragging of carousel is ended
+         */
         this._draggedCarousel$ = new Subject();
         /**
          * Current settings for the carousel.
@@ -282,8 +284,8 @@ class CarouselService {
             items: 0
         };
         /**
-       * Initial data for setting classes to element .owl-carousel
-       */
+         * Initial data for setting classes to element .owl-carousel
+         */
         this.owlDOMData = {
             rtl: false,
             isResponsive: false,
@@ -295,8 +297,8 @@ class CarouselService {
             isTouchDragable: false
         };
         /**
-       * Initial data of .owl-stage
-       */
+         * Initial data of .owl-stage
+         */
         this.stageData = {
             transform: 'translate3d(0px,0px,0px)',
             transition: '0s',
@@ -308,44 +310,45 @@ class CarouselService {
          * All real items.
          */
         this._items = []; // is equal to this.slides
+        // is equal to this.slides
         /**
-       * Array with width of every slide.
-       */
+         * Array with width of every slide.
+         */
         this._widths = [];
         /**
-       * Currently suppressed events to prevent them from beeing retriggered.
-       */
+         * Currently suppressed events to prevent them from beeing retriggered.
+         */
         this._supress = {};
         /**
          * References to the running plugins of this carousel.
          */
         this._plugins = {};
         /**
-       * Absolute current position.
-       */
+         * Absolute current position.
+         */
         this._current = null;
         /**
-       * All cloned items.
-       */
+         * All cloned items.
+         */
         this._clones = [];
         /**
          * Merge values of all items.
-         * @todo Maybe this could be part of a plugin.
+         * \@todo Maybe this could be part of a plugin.
          */
         this._mergers = [];
         /**
-       * Animation speed in milliseconds.
-       */
+         * Animation speed in milliseconds.
+         */
         this._speed = null;
         /**
-       * Coordinates of all items in pixel.
-       * @todo The name of this member is missleading.
-       */
+         * Coordinates of all items in pixel.
+         * \@todo The name of this member is missleading.
+         */
         this._coordinates = [];
         /**
-       * Current breakpoint.
-       * @todo Real media queries would be nice.
-       */
+         * Current breakpoint.
+         * \@todo Real media queries would be nice.
+         */
         this._breakpoint = null;
         /**
          * Prefix for id of cloned slides
@@ -372,7 +375,7 @@ class CarouselService {
         };
         /**
          * Ordered list of workers for the update process.
-       */
+         */
         this._pipe = [
             // {
             //   filter: ['width', 'settings'],
@@ -736,7 +739,6 @@ class CarouselService {
     }
     /**
      * Checks whether user's option are set properly. Cheking is based on typings;
-     * @private
      * @param {?} options options set by user
      * @param {?} configOptions default options
      * @return {?} checked and modified (if it's needed) user's options
@@ -799,7 +801,6 @@ class CarouselService {
     }
     /**
      * Checks option items set by user and if it bigger than number of slides then returns number of slides
-     * @private
      * @param {?} items option items set by user
      * @return {?} right number of items
      */
@@ -840,16 +841,16 @@ class CarouselService {
         this._defineSlidesData();
         this.setOptions(options);
         this.settings = Object.assign({}, this._options);
-        this.setViewportItemsN();
+        this.setOptionsForViewport();
         this._trigger('change', { property: { name: 'settings', value: this.settings } });
         this.invalidate('settings'); // must be call of this function;
         this._trigger('changed', { property: { name: 'settings', value: this.settings } });
     }
     /**
-     * Set number of items for current viewport
+     * Set options for current viewport
      * @return {?}
      */
-    setViewportItemsN() {
+    setOptionsForViewport() {
         /** @type {?} */
         const viewport = this._width;
         /** @type {?} */
@@ -870,12 +871,22 @@ class CarouselService {
                 }
             }
         }
-        this.settings = Object.assign({}, this.settings, { items: this._validateItems(overwrites[match].items) });
+        this.settings = Object.assign({}, this._options, overwrites[match], { items: (overwrites[match] && overwrites[match].items) ? this._validateItems(overwrites[match].items) : this._options.items });
         // if (typeof this.settings.stagePadding === 'function') {
         // 	this.settings.stagePadding = this.settings.stagePadding();
         // }
         delete this.settings.responsive;
         this.owlDOMData.isResponsive = true;
+        this.owlDOMData.isMouseDragable = this.settings.mouseDrag;
+        this.owlDOMData.isTouchDragable = this.settings.touchDrag;
+        /** @type {?} */
+        const mergers = [];
+        this._items.forEach(item => {
+            /** @type {?} */
+            const mergeN = this.settings.merge ? item.dataMerge : 1;
+            mergers.push(mergeN);
+        });
+        this._mergers = mergers;
         this._breakpoint = match;
         this.invalidate('settings');
     }
@@ -923,7 +934,6 @@ class CarouselService {
     }
     /**
      * Updates option logic if necessery
-     * @private
      * @return {?}
      */
     _optionsLogic() {
@@ -983,7 +993,7 @@ class CarouselService {
         this.enter('refreshing');
         this._trigger('refresh');
         this._defineSlidesData();
-        this.setViewportItemsN();
+        this.setOptionsForViewport();
         this._optionsLogic();
         // this.$element.addClass(this.options.refreshClass);
         this.update();
@@ -1387,11 +1397,11 @@ class CarouselService {
         /** @type {?} */
         const even = odd + this._items.length;
         /** @type {?} */
-        const map = index => index % 2 === 0 ? even + index / 2 : odd - (index + 1) / 2;
+        const map$$1 = index => index % 2 === 0 ? even + index / 2 : odd - (index + 1) / 2;
         if (position === undefined) {
-            return this._clones.map((v, i) => map(i));
+            return this._clones.map((v, i) => map$$1(i));
         }
-        return this._clones.map((v, i) => v === position ? map(i) : null).filter(item => item);
+        return this._clones.map((v, i) => v === position ? map$$1(i) : null).filter(item => item);
     }
     /**
      * Sets the current animation speed.
@@ -1441,17 +1451,16 @@ class CarouselService {
     }
     /**
      * Calculates the speed for a translation.
-     * @private
      * @param {?} from The absolute position of the start item.
      * @param {?} to The absolute position of the target item.
      * @param {?=} factor [factor=undefined] - The time factor in milliseconds.
      * @return {?} The time in milliseconds for the translation.
      */
-    _duration(from, to, factor) {
+    _duration(from$$1, to, factor) {
         if (factor === 0) {
             return 0;
         }
-        return Math.min(Math.max(Math.abs(to - from), 1), 6) * Math.abs((+factor || this.settings.smartSpeed));
+        return Math.min(Math.max(Math.abs(to - from$$1), 1), 6) * Math.abs((+factor || this.settings.smartSpeed));
     }
     /**
      * Slides to the specified item.
@@ -1538,7 +1547,6 @@ class CarouselService {
     }
     /**
      * Gets viewport width.
-     * @private
      * @return {?} - The width in pixel.
      */
     _viewport() {
@@ -1562,7 +1570,6 @@ class CarouselService {
     }
     /**
      * Sets slidesData using this._items
-     * @private
      * @return {?}
      */
     _defineSlidesData() {
@@ -1619,7 +1626,6 @@ class CarouselService {
     }
     /**
      * Operators to calculate right-to-left and left-to-right.
-     * @private
      * @param {?} a - The left side operand.
      * @param {?} o - The operator.
      * @param {?} b - The right side operand.
@@ -1644,7 +1650,6 @@ class CarouselService {
     /**
      * Triggers a public event.
      * \@todo Remove `status`, `relatedTarget` should be used instead.
-     * @private
      * @param {?} name The event name.
      * @param {?=} data The event data.
      * @param {?=} namespace The event namespace.
@@ -1738,7 +1743,6 @@ class CarouselService {
     }
     /**
      * Suppresses events.
-     * @private
      * @param {?} events The events to suppress.
      * @return {?}
      */
@@ -1749,7 +1753,6 @@ class CarouselService {
     }
     /**
      * Releases suppressed events.
-     * @private
      * @param {?} events The events to release.
      * @return {?}
      */
@@ -1783,7 +1786,6 @@ class CarouselService {
     }
     /**
      * Determines if the input is a Number or something that can be coerced to a Number
-     * @private
      * @param {?} number The input to be tested
      * @return {?} An indication if the input is a Number or can be coerced to a Number
      */
@@ -1792,7 +1794,6 @@ class CarouselService {
     }
     /**
      * Determines whether value is number or boolean type
-     * @private
      * @param {?} value The input to be tested
      * @return {?} An indication if the input is a Number or can be coerced to a Number, or Boolean
      */
@@ -1801,7 +1802,6 @@ class CarouselService {
     }
     /**
      * Determines whether value is number or string type
-     * @private
      * @param {?} value The input to be tested
      * @return {?} An indication if the input is a Number or can be coerced to a Number, or String
      */
@@ -1810,7 +1810,6 @@ class CarouselService {
     }
     /**
      * Determines whether value is number or string type
-     * @private
      * @param {?} value The input to be tested
      * @return {?} An indication if the input is a Number or can be coerced to a Number, or String
      */
@@ -1834,18 +1833,19 @@ class CarouselService {
 CarouselService.decorators = [
     { type: Injectable }
 ];
+/** @nocollapse */
 CarouselService.ctorParameters = () => [
     { type: OwlLogger }
 ];
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 class NavigationService {
     /**
@@ -1940,7 +1940,6 @@ class NavigationService {
     }
     /**
      * Calculates internal states and updates prop _pages
-     * @private
      * @return {?}
      */
     _updateNavPages() {
@@ -2020,6 +2019,7 @@ class NavigationService {
                     this._dotsData.dots.push({
                         active: false,
                         id: `dot-${i + startI}`,
+                        innerContent: '',
                         showInnerContent: false
                     });
                 }
@@ -2042,7 +2042,6 @@ class NavigationService {
     }
     /**
      * Changes state of nav buttons (disabled, enabled)
-     * @private
      * @return {?}
      */
     _updateNavButtons() {
@@ -2060,12 +2059,14 @@ class NavigationService {
     }
     /**
      * Changes active dot if page becomes changed
-     * @private
      * @return {?}
      */
     _updateDots() {
         /** @type {?} */
         let curActiveDotI;
+        if (!this.carouselService.settings.dots) {
+            return;
+        }
         this._dotsData.dots.forEach(item => {
             if (item.active === true) {
                 item.active = false;
@@ -2079,7 +2080,6 @@ class NavigationService {
     }
     /**
      * Gets the current page position of the carousel.
-     * @private
      * @return {?} the current page position of the carousel
      */
     _current() {
@@ -2099,7 +2099,6 @@ class NavigationService {
     ;
     /**
      * Gets the current succesor/predecessor position.
-     * @private
      * @param {?} successor
      * @return {?} the current succesor/predecessor position
      */
@@ -2188,13 +2187,14 @@ class NavigationService {
 NavigationService.decorators = [
     { type: Injectable }
 ];
+/** @nocollapse */
 NavigationService.ctorParameters = () => [
     { type: CarouselService }
 ];
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 /**
  * Create a new injection token for injecting the window into a component.
@@ -2237,7 +2237,12 @@ function windowFactory(browserWindowRef, platformId) {
     if (isPlatformBrowser(platformId)) {
         return browserWindowRef.nativeWindow;
     }
-    return new Object();
+    /** @type {?} */
+    const obj = {
+        setTimeout: (func, time) => { },
+        clearTimeout: (a) => { }
+    };
+    return obj;
 }
 /**
  * Create a injectable provider for the WindowRef token that uses the BrowserWindowRef class.
@@ -2264,7 +2269,7 @@ const WINDOW_PROVIDERS = [browserWindowProvider, windowProvider];
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 /**
  * Create a new injection token for injecting the Document into a component.
@@ -2307,7 +2312,12 @@ function documentFactory(browserDocumentRef, platformId) {
     if (isPlatformBrowser(platformId)) {
         return browserDocumentRef.nativeDocument;
     }
-    return new Object();
+    /** @type {?} */
+    const doc = {
+        hidden: false,
+        visibilityState: 'visible'
+    };
+    return doc;
 }
 /**
  * Create a injectable provider for the DocumentRef token that uses the BrowserDocumentRef class.
@@ -2334,7 +2344,7 @@ const DOCUMENT_PROVIDERS = [browserDocumentProvider, documentProvider];
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 class AutoplayService {
     /**
@@ -2377,10 +2387,22 @@ class AutoplayService {
         const changedSettings$ = this.carouselService.getChangedState().pipe(tap(data => {
             this._handleChangeObservable(data);
         }));
+        /** @type {?} */
+        const resized$ = this.carouselService.getResizedState().pipe(tap(() => {
+            if (this.carouselService.settings.autoplay) {
+                this.play();
+            }
+            else {
+                this.stop();
+            }
+        }))
+        // original Autoplay Plugin has listeners on play.owl.core and stop.owl.core events.
+        // They are triggered by Video Plugin
+        ;
         // original Autoplay Plugin has listeners on play.owl.core and stop.owl.core events.
         // They are triggered by Video Plugin
         /** @type {?} */
-        const autoplayMerge$ = merge(initializedCarousel$, changedSettings$);
+        const autoplayMerge$ = merge(initializedCarousel$, changedSettings$, resized$);
         this.autoplaySubscription = autoplayMerge$.subscribe(() => { });
     }
     /**
@@ -2392,7 +2414,7 @@ class AutoplayService {
     play(timeout, speed) {
         if (this._paused) {
             this._paused = false;
-            this._setAutoPlayInterval();
+            this._setAutoPlayInterval(1);
         }
         if (this.carouselService.is('rotating')) {
             return;
@@ -2403,7 +2425,6 @@ class AutoplayService {
     ;
     /**
      * Gets a new timeout
-     * @private
      * @param {?=} timeout - The interval before the next animation starts.
      * @param {?=} speed - The animation speed for the animations.
      * @return {?}
@@ -2412,6 +2433,7 @@ class AutoplayService {
         if (this._timeout) {
             this.winRef.clearTimeout(this._timeout);
         }
+        this._isArtificialAutoplayTimeout = timeout ? true : false;
         return this.winRef.setTimeout(() => {
             if (this._paused || this.carouselService.is('busy') || this.carouselService.is('interacting') || this.docRef.hidden) {
                 return;
@@ -2422,11 +2444,11 @@ class AutoplayService {
     ;
     /**
      * Sets autoplay in motion.
-     * @private
+     * @param {?=} timeout
      * @return {?}
      */
-    _setAutoPlayInterval() {
-        this._timeout = this._getNextTimeout();
+    _setAutoPlayInterval(timeout) {
+        this._timeout = this._getNextTimeout(timeout);
     }
     ;
     /**
@@ -2437,6 +2459,7 @@ class AutoplayService {
         if (!this.carouselService.is('rotating')) {
             return;
         }
+        this._paused = true;
         this.winRef.clearTimeout(this._timeout);
         this.carouselService.leave('rotating');
     }
@@ -2454,7 +2477,6 @@ class AutoplayService {
     ;
     /**
      * Manages by autoplaying according to data passed by _changedSettingsCarousel$ Obsarvable
-     * @private
      * @param {?} data object with current position of carousel and type of change
      * @return {?}
      */
@@ -2475,6 +2497,13 @@ class AutoplayService {
         }
     }
     /**
+     * Starts autoplaying of the carousel in the case when user leaves the carousel before it starts translateing (moving)
+     * @return {?}
+     */
+    _playAfterTranslated() {
+        of('translated').pipe(switchMap(data => this.carouselService.getTranslatedState()), first(), filter(() => this._isArtificialAutoplayTimeout), tap(() => this._setAutoPlayInterval())).subscribe(() => { });
+    }
+    /**
      * Starts pausing
      * @return {?}
      */
@@ -2489,7 +2518,8 @@ class AutoplayService {
      */
     startPlayingMouseLeave() {
         if (this.carouselService.settings.autoplayHoverPause && this.carouselService.is('rotating')) {
-            this.pause();
+            this.play();
+            this._playAfterTranslated();
         }
     }
     /**
@@ -2498,13 +2528,15 @@ class AutoplayService {
      */
     startPlayingTouchEnd() {
         if (this.carouselService.settings.autoplayHoverPause && this.carouselService.is('rotating')) {
-            this.pause();
+            this.play();
+            this._playAfterTranslated();
         }
     }
 }
 AutoplayService.decorators = [
     { type: Injectable }
 ];
+/** @nocollapse */
 AutoplayService.ctorParameters = () => [
     { type: CarouselService },
     { type: undefined, decorators: [{ type: Inject, args: [WINDOW,] }] },
@@ -2513,7 +2545,7 @@ AutoplayService.ctorParameters = () => [
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 class LazyLoadService {
     /**
@@ -2549,7 +2581,6 @@ class LazyLoadService {
         this.lazyLoadSubscription = lazyLoadMerge$.subscribe(() => { });
     }
     /**
-     * @private
      * @param {?} data
      * @return {?}
      */
@@ -2589,7 +2620,6 @@ class LazyLoadService {
     }
     /**
      * Loads all resources of an item at the specified position.
-     * @private
      * @param {?} position - The absolute position of the item.
      * @return {?}
      */
@@ -2603,13 +2633,14 @@ class LazyLoadService {
 LazyLoadService.decorators = [
     { type: Injectable }
 ];
+/** @nocollapse */
 LazyLoadService.ctorParameters = () => [
     { type: CarouselService }
 ];
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 class AnimateService {
     /**
@@ -2669,7 +2700,6 @@ class AnimateService {
     }
     /**
      * Toggles the animation classes whenever an translations starts.
-     * @private
      * @return {?}
      */
     _swap() {
@@ -2739,13 +2769,14 @@ class AnimateService {
 AnimateService.decorators = [
     { type: Injectable }
 ];
+/** @nocollapse */
 AnimateService.ctorParameters = () => [
     { type: CarouselService }
 ];
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 class AutoHeightService {
     /**
@@ -2814,13 +2845,14 @@ class AutoHeightService {
 AutoHeightService.decorators = [
     { type: Injectable }
 ];
+/** @nocollapse */
 AutoHeightService.ctorParameters = () => [
     { type: CarouselService }
 ];
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 class HashService {
     /**
@@ -2833,6 +2865,16 @@ class HashService {
         this.route = route;
         this.router = router;
         this.spyDataStreams();
+        if (!this.route) {
+            this.route = (/** @type {?} */ ({
+                fragment: of('no route').pipe(take(1))
+            }));
+        }
+        if (!this.router) {
+            this.router = (/** @type {?} */ ({
+                navigate: (commands, extras) => { return; }
+            }));
+        }
     }
     /**
      * @return {?}
@@ -2894,15 +2936,16 @@ class HashService {
 HashService.decorators = [
     { type: Injectable }
 ];
+/** @nocollapse */
 HashService.ctorParameters = () => [
     { type: CarouselService },
-    { type: ActivatedRoute },
-    { type: Router }
+    { type: ActivatedRoute, decorators: [{ type: Optional }] },
+    { type: Router, decorators: [{ type: Optional }] }
 ];
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 /** @type {?} */
 let nextId = 0;
@@ -2959,6 +3002,7 @@ class CarouselSlideDirective {
 CarouselSlideDirective.decorators = [
     { type: Directive, args: [{ selector: 'ng-template[carouselSlide]' },] }
 ];
+/** @nocollapse */
 CarouselSlideDirective.ctorParameters = () => [
     { type: TemplateRef }
 ];
@@ -2986,8 +3030,9 @@ class CarouselComponent {
      * @param {?} autoHeightService
      * @param {?} hashService
      * @param {?} logger
+     * @param {?} docRef
      */
-    constructor(el, resizeService, carouselService, navigationService, autoplayService, lazyLoadService, animateService, autoHeightService, hashService, logger) {
+    constructor(el, resizeService, carouselService, navigationService, autoplayService, lazyLoadService, animateService, autoHeightService, hashService, logger, docRef) {
         this.el = el;
         this.resizeService = resizeService;
         this.carouselService = carouselService;
@@ -3001,6 +3046,7 @@ class CarouselComponent {
         this.translated = new EventEmitter();
         this.dragging = new EventEmitter();
         this.change = new EventEmitter();
+        this.changed = new EventEmitter();
         this.initialized = new EventEmitter();
         /**
          *  Data of every slide
@@ -3010,7 +3056,27 @@ class CarouselComponent {
          * Shows whether carousel is loaded of not.
          */
         this.carouselLoaded = false;
+        this.docRef = (/** @type {?} */ (docRef));
     }
+    /**
+     * @param {?} ev
+     * @return {?}
+     */
+    onVisibilityChange(ev) {
+        if (!this.carouselService.settings.autoplay)
+            return;
+        switch (this.docRef.visibilityState) {
+            case 'visible':
+                this.autoplayService.play();
+                break;
+            case 'hidden':
+                this.autoplayService.stop();
+                break;
+            default:
+                break;
+        }
+    }
+    ;
     /**
      * @return {?}
      */
@@ -3089,18 +3155,54 @@ class CarouselComponent {
             this.change.emit(this.slidesOutputData);
             // this.slidesOutputData = {};
         }));
+        this._changedCarousel$ = this.carouselService.getChangeState().pipe(switchMap(value => {
+            /** @type {?} */
+            const changedPosition = of(value).pipe(filter(() => value.property.name === 'position'), switchMap(() => from(this.slidesData)), skip(value.property.value), take(this.carouselService.settings.items), map(slide => {
+                /** @type {?} */
+                const clonedIdPrefix = this.carouselService.clonedIdPrefix;
+                /** @type {?} */
+                const id = slide.id.indexOf(clonedIdPrefix) >= 0 ? slide.id.slice(clonedIdPrefix.length) : slide.id;
+                return Object.assign({}, slide, { id: id, isActive: true });
+            }), toArray(), map(slides => {
+                return {
+                    slides: slides,
+                    startPosition: this.carouselService.relative(value.property.value)
+                };
+            }));
+            // const changedSetting: Observable<SlidesOutputData> = of(value).pipe(
+            //   filter(() => value.property.name === 'settings'),
+            //   map(() => {
+            //     return {
+            //       slides: [],
+            //       startPosition: this.carouselService.relative(value.property.value)
+            //     }
+            //   })
+            // )
+            return merge(changedPosition);
+        }), tap(slidesData => {
+            this.gatherTranslatedData();
+            this.changed.emit(slidesData.slides.length ? slidesData : this.slidesOutputData);
+            // console.log(this.slidesOutputData);
+            // this.slidesOutputData = {};
+        }));
         this._draggingCarousel$ = this.carouselService.getDragState().pipe(tap(() => {
             this.gatherTranslatedData();
             this.dragging.emit({ dragging: true, data: this.slidesOutputData });
-        }), switchMap(() => this.carouselService.getTranslatedState().pipe(first(), tap(() => {
+        }), switchMap(() => this.carouselService.getDraggedState().pipe(map(() => !!this.carouselService.is('animating')))), switchMap(anim => {
+            if (anim) {
+                return this.carouselService.getTranslatedState().pipe(first());
+            }
+            else {
+                return of('not animating');
+            }
+        }), tap(() => {
             this.dragging.emit({ dragging: false, data: this.slidesOutputData });
-        }))));
-        this._carouselMerge$ = merge(this._viewCurSettings$, this._translatedCarousel$, this._draggingCarousel$, this._changeCarousel$, this._initializedCarousel$);
+        }));
+        this._carouselMerge$ = merge(this._viewCurSettings$, this._translatedCarousel$, this._draggingCarousel$, this._changeCarousel$, this._changedCarousel$, this._initializedCarousel$);
         this._allObservSubscription = this._carouselMerge$.subscribe(() => { });
     }
     /**
      * Init subscription to resize event and attaches handler for this event
-     * @private
      * @return {?}
      */
     _winResizeWatcher() {
@@ -3125,7 +3227,7 @@ class CarouselComponent {
      * @return {?}
      */
     next() {
-        if (!this.carouselLoaded || (this.navData && this.navData.disabled))
+        if (!this.carouselLoaded)
             return;
         this.navigationService.next(this.carouselService.settings.navSpeed);
     }
@@ -3134,7 +3236,7 @@ class CarouselComponent {
      * @return {?}
      */
     prev() {
-        if (!this.carouselLoaded || (this.navData && this.navData.disabled))
+        if (!this.carouselLoaded)
             return;
         this.navigationService.prev(this.carouselService.settings.navSpeed);
     }
@@ -3154,7 +3256,8 @@ class CarouselComponent {
      * @return {?}
      */
     to(id) {
-        if (!this.carouselLoaded || (this.navData && this.navData.disabled) || (this.dotsData && this.dotsData.disabled))
+        // if (!this.carouselLoaded || ((this.navData && this.navData.disabled) && (this.dotsData && this.dotsData.disabled))) return;
+        if (!this.carouselLoaded)
             return;
         this.navigationService.toSlideById(id);
     }
@@ -3254,6 +3357,7 @@ CarouselComponent.decorators = [
                 styles: [`.owl-theme { display: block; }`]
             }] }
 ];
+/** @nocollapse */
 CarouselComponent.ctorParameters = () => [
     { type: ElementRef },
     { type: ResizeService },
@@ -3264,25 +3368,28 @@ CarouselComponent.ctorParameters = () => [
     { type: AnimateService },
     { type: AutoHeightService },
     { type: HashService },
-    { type: OwlLogger }
+    { type: OwlLogger },
+    { type: undefined, decorators: [{ type: Inject, args: [DOCUMENT,] }] }
 ];
 CarouselComponent.propDecorators = {
     slides: [{ type: ContentChildren, args: [CarouselSlideDirective,] }],
     translated: [{ type: Output }],
     dragging: [{ type: Output }],
     change: [{ type: Output }],
+    changed: [{ type: Output }],
     initialized: [{ type: Output }],
-    options: [{ type: Input }]
+    options: [{ type: Input }],
+    onVisibilityChange: [{ type: HostListener, args: ['document:visibilitychange', ['$event'],] }]
 };
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 class StageComponent {
     /**
@@ -3405,7 +3512,6 @@ class StageComponent {
     /**
      * Handles `touchstart` and `mousedown` events.
      * \@todo Horizontal swipe threshold as option / #261
-     * @private
      * @param {?} event - The event arguments.
      * @return {?}
      */
@@ -3421,7 +3527,6 @@ class StageComponent {
         this._drag.stage.start = stage;
         this._drag.stage.current = stage;
         this._drag.pointer = this._pointer(event);
-        this._drag.active = true;
         this.listenerMouseUp = this.renderer.listen(document, 'mouseup', this.bindOnDragEnd);
         this.listenerTouchEnd = this.renderer.listen(document, 'touchend', this.bindOnDragEnd);
         this.zone.runOutsideAngular(() => {
@@ -3431,24 +3536,23 @@ class StageComponent {
     }
     /**
      * Attaches listeners to `touchmove` and `mousemove` events; initiates updating carousel after starting dragging
-     * @private
      * @param {?} event event objech of mouse or touch event
      * @return {?}
      */
     _oneMouseTouchMove(event) {
-        if (!this._drag.active)
-            return false;
         /** @type {?} */
         const delta = this._difference(this._drag.pointer, this._pointer(event));
         if (this.listenerATag) {
             this.listenerATag();
         }
-        this.listenerOneMouseMove();
-        this.listenerOneTouchMove();
-        if (Math.abs(delta.x) < Math.abs(delta.y) && this._is('valid')) {
-            this._drag.active = false;
+        if (Math.abs(delta.x) < 3 && Math.abs(delta.y) < 3 && this._is('valid')) {
             return;
         }
+        if ((Math.abs(delta.x) < 3 && Math.abs(delta.x) < Math.abs(delta.y)) && this._is('valid')) {
+            return;
+        }
+        this.listenerOneMouseMove();
+        this.listenerOneTouchMove();
         this._drag.moving = true;
         this.blockClickAnchorInDragging(event);
         this.listenerMouseMove = this.renderer.listen(document, 'mousemove', this.bindOnDragMove);
@@ -3460,7 +3564,6 @@ class StageComponent {
     }
     /**
      * Attaches handler to HTMLAnchorElement for preventing click while carousel is being dragged
-     * @private
      * @param {?} event event object
      * @return {?}
      */
@@ -3477,13 +3580,10 @@ class StageComponent {
     /**
      * Handles the `touchmove` and `mousemove` events.
      * \@todo #261
-     * @private
      * @param {?} event - The event arguments.
      * @return {?}
      */
     _onDragMove(event) {
-        if (!this._drag.active)
-            return false;
         /** @type {?} */
         let stage;
         /** @type {?} */
@@ -3499,7 +3599,6 @@ class StageComponent {
     ;
     /**
      * Moves .owl-stage left-right
-     * @private
      * @param {?} coordinate coordinate to be set to .owl-stage
      * @return {?}
      */
@@ -3510,12 +3609,13 @@ class StageComponent {
     /**
      * Handles the `touchend` and `mouseup` events.
      * \@todo #261 / Threshold for click event
-     * @private
      * @param {?} event - The event arguments.
      * @return {?}
      */
     _onDragEnd(event) {
         this.carouselService.owlDOMData.isGrab = false;
+        this.listenerOneMouseMove();
+        this.listenerOneTouchMove();
         if (this._drag.moving) {
             this.renderer.setStyle(this.el.nativeElement.children[0], 'transform', ``);
             this.renderer.setStyle(this.el.nativeElement.children[0], 'transition', this.carouselService.speed(+this.carouselService.settings.dragEndSpeed || this.carouselService.settings.smartSpeed) / 1000 + 's');
@@ -3542,7 +3642,6 @@ class StageComponent {
     ;
     /**
      * Prepares data for dragging carousel. It starts after firing `touchstart` and `mousedown` events.
-     * @private
      * @param {?} event - The event arguments.
      * @return {?} stage - object with 'x' and 'y' coordinates of .owl-stage
      */
@@ -3551,7 +3650,6 @@ class StageComponent {
     }
     /**
      * Finishes dragging
-     * @private
      * @param {?} event object event of 'mouseUp' of 'touchend' events
      * @return {?}
      */
@@ -3560,7 +3658,6 @@ class StageComponent {
     }
     /**
      * Gets unified pointer coordinates from event.
-     * @private
      * @param {?} event The `mousedown` or `touchstart` event.
      * @return {?} Contains `x` and `y` coordinates of current pointer position.
      */
@@ -3569,7 +3666,6 @@ class StageComponent {
     }
     /**
      * Gets the difference of two vectors.
-     * @private
      * @param {?} firstC
      * @param {?} second
      * @return {?} The difference.
@@ -3579,7 +3675,6 @@ class StageComponent {
     }
     /**
      * Checks whether the carousel is in a specific state or not.
-     * @private
      * @param {?} specificState The state to check.
      * @return {?} The flag which indicates if the carousel is busy.
      */
@@ -3588,7 +3683,6 @@ class StageComponent {
     }
     /**
      * Enters a state.
-     * @private
      * @param {?} name The state name.
      * @return {?}
      */
@@ -3597,7 +3691,6 @@ class StageComponent {
     }
     /**
      * Sends all data needed for View.
-     * @private
      * @return {?}
      */
     _sendChanges() {
@@ -3612,7 +3705,6 @@ class StageComponent {
     }
     /**
      * Enters into a 'dragging' state
-     * @private
      * @return {?}
      */
     _enterDragging() {
@@ -3635,14 +3727,14 @@ StageComponent.decorators = [
       <div class="owl-stage" [ngStyle]="{'width': stageData.width + 'px',
                                         'transform': stageData.transform,
                                         'transition': stageData.transition,
-                                        'padding-left': stageData.paddingL + 'px',
-                                        'padding-right': stageData.paddingR + 'px' }"
+                                        'padding-left': stageData.paddingL ? stageData.paddingL + 'px' : '',
+                                        'padding-right': stageData.paddingR ? stageData.paddingR + 'px' : '' }"
           (transitionend)="onTransitionEnd()">
         <ng-container *ngFor="let slide of slidesData; let i = index">
           <div class="owl-item" [ngClass]="slide.classes"
                                 [ngStyle]="{'width': slide.width + 'px',
-                                            'margin-left': slide.marginL + 'px',
-                                            'margin-right': slide.marginR + 'px',
+                                            'margin-left': slide.marginL ? slide.marginL + 'px' : '',
+                                            'margin-right': slide.marginR ? slide.marginR + 'px' : '',
                                             'left': slide.left}"
                                 (animationend)="clear(slide.id)"
                                 [@autoHeight]="slide.heightState">
@@ -3668,6 +3760,7 @@ StageComponent.decorators = [
                 ]
             }] }
 ];
+/** @nocollapse */
 StageComponent.ctorParameters = () => [
     { type: NgZone },
     { type: ElementRef },
@@ -3688,7 +3781,7 @@ StageComponent.propDecorators = {
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 class OwlRouterLinkDirective {
     /**
@@ -3762,6 +3855,7 @@ class OwlRouterLinkDirective {
 OwlRouterLinkDirective.decorators = [
     { type: Directive, args: [{ selector: ':not(a)[owlRouterLink]' },] }
 ];
+/** @nocollapse */
 OwlRouterLinkDirective.ctorParameters = () => [
     { type: Router },
     { type: ActivatedRoute },
@@ -3867,7 +3961,6 @@ class OwlRouterLinkWithHrefDirective {
         return false;
     }
     /**
-     * @private
      * @return {?}
      */
     updateTargetUrlAndHref() {
@@ -3890,6 +3983,7 @@ class OwlRouterLinkWithHrefDirective {
 OwlRouterLinkWithHrefDirective.decorators = [
     { type: Directive, args: [{ selector: 'a[owlRouterLink]' },] }
 ];
+/** @nocollapse */
 OwlRouterLinkWithHrefDirective.ctorParameters = () => [
     { type: Router },
     { type: ActivatedRoute },
@@ -3919,18 +4013,14 @@ function attrBoolValue(s) {
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
-/** @type {?} */
-const routes = [];
 class CarouselModule {
 }
 CarouselModule.decorators = [
     { type: NgModule, args: [{
                 imports: [
                     CommonModule,
-                    // BrowserAnimationsModule, // there's an issue with this import while using lazy loading of module consuming this library. I don't remove it because it could be needed during future enhancement of this lib.
-                    RouterModule.forChild(routes)
                 ],
                 declarations: [CarouselComponent, CarouselSlideDirective, StageComponent, OwlRouterLinkDirective, OwlRouterLinkWithHrefDirective],
                 exports: [CarouselComponent, CarouselSlideDirective, OwlRouterLinkDirective, OwlRouterLinkWithHrefDirective],
@@ -3940,17 +4030,17 @@ CarouselModule.decorators = [
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 
 export { CarouselModule, CarouselComponent, CarouselSlideDirective, SlidesOutputData, OwlRouterLinkDirective, OwlRouterLinkWithHrefDirective, StageComponent as ɵx, AnimateService as ɵt, AutoHeightService as ɵu, AutoplayService as ɵd, CarouselService as ɵb, BrowserDocumentRef as ɵn, DOCUMENT as ɵl, DOCUMENT_PROVIDERS as ɵr, DocumentRef as ɵm, browserDocumentProvider as ɵp, documentFactory as ɵo, documentProvider as ɵq, HashService as ɵv, LazyLoadService as ɵs, OwlLogger as ɵc, NavigationService as ɵa, ResizeService as ɵw, BrowserWindowRef as ɵg, WINDOW as ɵe, WINDOW_PROVIDERS as ɵk, WindowRef as ɵf, browserWindowProvider as ɵi, windowFactory as ɵh, windowProvider as ɵj };
