@@ -1,7 +1,8 @@
 import { Injectable, Inject, OnDestroy } from '@angular/core';
-import { Subscription, Observable, merge } from 'rxjs';
+import { Subscription, Observable, merge, of } from 'rxjs';
+import { tap, switchMap, first, filter } from 'rxjs/operators';
+
 import { CarouselService } from './carousel.service';
-import { tap } from 'rxjs/operators';
 import { WINDOW } from './window-ref.service';
 import { DOCUMENT } from './document-ref.service';
 
@@ -21,6 +22,11 @@ export class AutoplayService implements OnDestroy{
    * Indicates whenever the autoplay is paused.
    */
   private _paused = false;
+
+  /**
+   * Shows whether the code (the plugin) changed the option 'AutoplayTimeout' for own needs
+   */
+  private _isArtificialAutoplayTimeout: boolean;
 
   private winRef: Window;
   private docRef: Document;
@@ -45,7 +51,7 @@ export class AutoplayService implements OnDestroy{
     const initializedCarousel$: Observable<string> = this.carouselService.getInitializedState().pipe(
       tap(() => {
         if (this.carouselService.settings.autoplay) {
-					this.play();
+          this.play();
 				}
       })
     );
@@ -56,10 +62,20 @@ export class AutoplayService implements OnDestroy{
       })
     );
 
+    const resized$: Observable<any> = this.carouselService.getResizedState().pipe(
+      tap(() => {
+        if (this.carouselService.settings.autoplay) {
+          this.play();
+				} else {
+          this.stop();
+        }
+      })
+    )
+
     // original Autoplay Plugin has listeners on play.owl.core and stop.owl.core events.
     // They are triggered by Video Plugin
 
-    const autoplayMerge$: Observable<string> = merge(initializedCarousel$, changedSettings$);
+    const autoplayMerge$: Observable<string> = merge(initializedCarousel$, changedSettings$, resized$);
     this.autoplaySubscription = autoplayMerge$.subscribe(
       () => {}
     );
@@ -73,7 +89,7 @@ export class AutoplayService implements OnDestroy{
 	play(timeout?: number, speed?: number) {
     if (this._paused) {
 			this._paused = false;
-			this._setAutoPlayInterval();
+			this._setAutoPlayInterval(1);
     }
 
 		if (this.carouselService.is('rotating')) {
@@ -94,7 +110,10 @@ export class AutoplayService implements OnDestroy{
 	private _getNextTimeout(timeout?: number, speed?: number): number {
 		if ( this._timeout ) {
 			this.winRef.clearTimeout(this._timeout);
-		}
+    }
+
+    this._isArtificialAutoplayTimeout = timeout ? true : false;
+
 		return this.winRef.setTimeout(() =>{
       if (this._paused || this.carouselService.is('busy') || this.carouselService.is('interacting') || this.docRef.hidden) {
 				return;
@@ -106,8 +125,8 @@ export class AutoplayService implements OnDestroy{
   /**
 	 * Sets autoplay in motion.
 	 */
-	private _setAutoPlayInterval() {
-		this._timeout = this._getNextTimeout();
+  private _setAutoPlayInterval(timeout?: number) {
+		this._timeout = this._getNextTimeout(timeout);
 	};
 
 	/**
@@ -117,6 +136,7 @@ export class AutoplayService implements OnDestroy{
 		if (!this.carouselService.is('rotating')) {
 			return;
 		}
+    this._paused = true;
 
 		this.winRef.clearTimeout(this._timeout);
 		this.carouselService.leave('rotating');
@@ -137,7 +157,7 @@ export class AutoplayService implements OnDestroy{
    * Manages by autoplaying according to data passed by _changedSettingsCarousel$ Obsarvable
    * @param data object with current position of carousel and type of change
    */
-  private _handleChangeObservable(data) {
+  private _handleChangeObservable(data: any) {
     if (data.property.name === 'settings') {
       if (this.carouselService.settings.autoplay) {
         this.play();
@@ -150,6 +170,18 @@ export class AutoplayService implements OnDestroy{
         this._setAutoPlayInterval();
       }
     }
+  }
+
+  /**
+   * Starts autoplaying of the carousel in the case when user leaves the carousel before it starts translateing (moving)
+   */
+  private _playAfterTranslated() {
+    of('translated').pipe(
+      switchMap(data => this.carouselService.getTranslatedState()),
+      first(),
+      filter(() => this._isArtificialAutoplayTimeout),
+      tap(() => this._setAutoPlayInterval())
+    ).subscribe(() => { });
   }
 
   /**
@@ -166,7 +198,8 @@ export class AutoplayService implements OnDestroy{
    */
   startPlayingMouseLeave() {
     if (this.carouselService.settings.autoplayHoverPause && this.carouselService.is('rotating')) {
-      this.pause();
+      this.play();
+      this._playAfterTranslated();
     }
   }
 
@@ -175,7 +208,8 @@ export class AutoplayService implements OnDestroy{
    */
   startPlayingTouchEnd() {
     if (this.carouselService.settings.autoplayHoverPause && this.carouselService.is('rotating')) {
-      this.pause();
+      this.play();
+      this._playAfterTranslated();
     }
   }
 }
