@@ -48,7 +48,11 @@ export function documentFactory(
   if (isPlatformBrowser(platformId)) {
     return browserDocumentRef.nativeDocument;
   }
-  return new Object();
+  const doc = {
+    hidden: false,
+    visibilityState: 'visible'
+  }
+  return doc;
 }
 
 /**
