@@ -4,80 +4,6 @@
     (factory((global['ngx-owl-carousel-o'] = {}),global.ng.platformBrowser,global.rxjs,global.rxjs.operators,global.ng.animations,global.ng.common,global.ng.core,global.ng.router));
 }(this, (function (exports,platformBrowser,rxjs,operators,animations,common,core,router) { 'use strict';
 
-    /**
-     * @fileoverview added by tsickle
-     * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
-     */
-    var ResizeService = (function () {
-        function ResizeService(eventManager) {
-            this.eventManager = eventManager;
-            this.resizeSubject = new rxjs.Subject();
-            this.eventManager.addGlobalEventListener('window', 'resize', this.onResize.bind(this));
-            this.eventManager.addGlobalEventListener('window', 'onload', this.onLoaded.bind(this));
-        }
-        Object.defineProperty(ResizeService.prototype, "onResize$", {
-            /**
-             * Makes resizeSubject become Observable
-             * @returns Observable of resizeSubject
-             */
-            get: /**
-             * Makes resizeSubject become Observable
-             * @return {?} Observable of resizeSubject
-             */ function () {
-                return this.resizeSubject.asObservable();
-            },
-            enumerable: true,
-            configurable: true
-        });
-        /**
-         * Handler of 'resize' event. Passes data throw resizeSubject
-         * @param event Event Object of 'resize' event
-         */
-        /**
-         * Handler of 'resize' event. Passes data throw resizeSubject
-         * @private
-         * @param {?} event Event Object of 'resize' event
-         * @return {?}
-         */
-        ResizeService.prototype.onResize = /**
-         * Handler of 'resize' event. Passes data throw resizeSubject
-         * @private
-         * @param {?} event Event Object of 'resize' event
-         * @return {?}
-         */
-            function (event) {
-                this.resizeSubject.next(((event.target)));
-            };
-        /**
-         * Handler of 'onload' event. Defines the width of window
-         * @param event Event Object of 'onload' event
-         */
-        /**
-         * Handler of 'onload' event. Defines the width of window
-         * @private
-         * @param {?} event Event Object of 'onload' event
-         * @return {?}
-         */
-        ResizeService.prototype.onLoaded = /**
-         * Handler of 'onload' event. Defines the width of window
-         * @private
-         * @param {?} event Event Object of 'onload' event
-         * @return {?}
-         */
-            function (event) {
-                this.windowWidth = ((event.target));
-            };
-        ResizeService.decorators = [
-            { type: core.Injectable }
-        ];
-        ResizeService.ctorParameters = function () {
-            return [
-                { type: platformBrowser.EventManager }
-            ];
-        };
-        return ResizeService;
-    }());
-
     /*! *****************************************************************************
     Copyright (c) Microsoft Corporation. All rights reserved.
     Licensed under the Apache License, Version 2.0 (the "License"); you may not use
@@ -150,14 +76,85 @@
 
     /**
      * @fileoverview added by tsickle
-     * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+     * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
+     */
+    var ResizeService = /** @class */ (function () {
+        function ResizeService(eventManager) {
+            this.eventManager = eventManager;
+            this.resizeSubject = new rxjs.Subject();
+            this.eventManager.addGlobalEventListener('window', 'resize', this.onResize.bind(this));
+            this.eventManager.addGlobalEventListener('window', 'onload', this.onLoaded.bind(this));
+        }
+        Object.defineProperty(ResizeService.prototype, "onResize$", {
+            /**
+             * Makes resizeSubject become Observable
+             * @returns Observable of resizeSubject
+             */
+            get: /**
+             * Makes resizeSubject become Observable
+             * @return {?} Observable of resizeSubject
+             */ function () {
+                return this.resizeSubject.asObservable();
+            },
+            enumerable: true,
+            configurable: true
+        });
+        /**
+         * Handler of 'resize' event. Passes data throw resizeSubject
+         * @param event Event Object of 'resize' event
+         */
+        /**
+         * Handler of 'resize' event. Passes data throw resizeSubject
+         * @param {?} event Event Object of 'resize' event
+         * @return {?}
+         */
+        ResizeService.prototype.onResize = /**
+         * Handler of 'resize' event. Passes data throw resizeSubject
+         * @param {?} event Event Object of 'resize' event
+         * @return {?}
+         */
+            function (event) {
+                this.resizeSubject.next(( /** @type {?} */(event.target)));
+            };
+        /**
+         * Handler of 'onload' event. Defines the width of window
+         * @param event Event Object of 'onload' event
+         */
+        /**
+         * Handler of 'onload' event. Defines the width of window
+         * @param {?} event Event Object of 'onload' event
+         * @return {?}
+         */
+        ResizeService.prototype.onLoaded = /**
+         * Handler of 'onload' event. Defines the width of window
+         * @param {?} event Event Object of 'onload' event
+         * @return {?}
+         */
+            function (event) {
+                this.windowWidth = ( /** @type {?} */(event.target));
+            };
+        ResizeService.decorators = [
+            { type: core.Injectable }
+        ];
+        /** @nocollapse */
+        ResizeService.ctorParameters = function () {
+            return [
+                { type: platformBrowser.EventManager }
+            ];
+        };
+        return ResizeService;
+    }());
+
+    /**
+     * @fileoverview added by tsickle
+     * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
      */
     /**
      * Defaults value of options
      */
     var /**
      * Defaults value of options
-     */ OwlCarouselOConfig = (function () {
+     */ OwlCarouselOConfig = /** @class */ (function () {
         function OwlCarouselOConfig() {
             this.items = 3;
             this.loop = false;
@@ -184,6 +181,7 @@
             this.navText = ['prev', 'next'];
             this.navSpeed = false;
             this.slideBy = 1; // stage moves on 1 width of slide; if slideBy = 2, stage moves on 2 widths of slide
+            // stage moves on 1 width of slide; if slideBy = 2, stage moves on 2 widths of slide
             this.dots = true;
             this.dotsEach = false;
             this.dotsData = false;
@@ -215,7 +213,7 @@
      * we can't read types from OwlOptions in javascript because of props have undefined value and types of those props are used for validating inputs
      * class below is copy of OwlOptions but its all props have string value showing certain type;
      * this is class is being used just in method _validateOptions() of CarouselService;
-     */ OwlOptionsMockedTypes = (function () {
+     */ OwlOptionsMockedTypes = /** @class */ (function () {
         function OwlOptionsMockedTypes() {
             this.items = 'number';
             this.loop = 'boolean';
@@ -242,6 +240,7 @@
             this.navText = 'string[]';
             this.navSpeed = 'number|boolean';
             this.slideBy = 'number|string'; // stage moves on 1 width of slide; if slideBy = 2, stage moves on 2 widths of slide
+            // stage moves on 1 width of slide; if slideBy = 2, stage moves on 2 widths of slide
             this.dots = 'boolean';
             this.dotsEach = 'number|boolean';
             this.dotsData = 'boolean';
@@ -267,9 +266,9 @@
 
     /**
      * @fileoverview added by tsickle
-     * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+     * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
      */
-    var OwlLogger = (function () {
+    var OwlLogger = /** @class */ (function () {
         function OwlLogger(errorHandler) {
             this.errorHandler = errorHandler;
         }
@@ -323,6 +322,7 @@
         OwlLogger.decorators = [
             { type: core.Injectable }
         ];
+        /** @nocollapse */
         OwlLogger.ctorParameters = function () {
             return [
                 { type: core.ErrorHandler }
@@ -333,7 +333,7 @@
 
     /**
      * @fileoverview added by tsickle
-     * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+     * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
      */
     /** @enum {string} */
     var Type = {
@@ -346,57 +346,57 @@
         Inner: 'inner',
         Outer: 'outer',
     };
-    var CarouselService = (function () {
+    var CarouselService = /** @class */ (function () {
         function CarouselService(logger) {
             var _this = this;
             this.logger = logger;
             /**
-           * Subject for passing data needed for managing View
-           */
+             * Subject for passing data needed for managing View
+             */
             this._viewSettingsShipper$ = new rxjs.Subject();
             /**
-           * Subject for notification when the carousel got initializes
-           */
+             * Subject for notification when the carousel got initializes
+             */
             this._initializedCarousel$ = new rxjs.Subject();
             /**
-           * Subject for notification when the carousel's settings start changinf
-           */
+             * Subject for notification when the carousel's settings start changinf
+             */
             this._changeSettingsCarousel$ = new rxjs.Subject();
             /**
-           * Subject for notification when the carousel's settings have changed
-           */
+             * Subject for notification when the carousel's settings have changed
+             */
             this._changedSettingsCarousel$ = new rxjs.Subject();
             /**
-           * Subject for notification when the carousel starts translating or moving
-           */
+             * Subject for notification when the carousel starts translating or moving
+             */
             this._translateCarousel$ = new rxjs.Subject();
             /**
-           * Subject for notification when the carousel stopped translating or moving
-           */
+             * Subject for notification when the carousel stopped translating or moving
+             */
             this._translatedCarousel$ = new rxjs.Subject();
             /**
-           * Subject for notification when the carousel's rebuilding caused by 'resize' event starts
-           */
+             * Subject for notification when the carousel's rebuilding caused by 'resize' event starts
+             */
             this._resizeCarousel$ = new rxjs.Subject();
             /**
-           * Subject for notification  when the carousel's rebuilding caused by 'resize' event is ended
-           */
+             * Subject for notification  when the carousel's rebuilding caused by 'resize' event is ended
+             */
             this._resizedCarousel$ = new rxjs.Subject();
             /**
-           * Subject for notification when the refresh of carousel starts
-           */
+             * Subject for notification when the refresh of carousel starts
+             */
             this._refreshCarousel$ = new rxjs.Subject();
             /**
-           * Subject for notification when the refresh of carousel is ended
-           */
+             * Subject for notification when the refresh of carousel is ended
+             */
             this._refreshedCarousel$ = new rxjs.Subject();
             /**
-           * Subject for notification when the dragging of carousel starts
-           */
+             * Subject for notification when the dragging of carousel starts
+             */
             this._dragCarousel$ = new rxjs.Subject();
             /**
-           * Subject for notification when the dragging of carousel is ended
-           */
+             * Subject for notification when the dragging of carousel is ended
+             */
             this._draggedCarousel$ = new rxjs.Subject();
             /**
              * Current settings for the carousel.
@@ -405,8 +405,8 @@
                 items: 0
             };
             /**
-           * Initial data for setting classes to element .owl-carousel
-           */
+             * Initial data for setting classes to element .owl-carousel
+             */
             this.owlDOMData = {
                 rtl: false,
                 isResponsive: false,
@@ -418,8 +418,8 @@
                 isTouchDragable: false
             };
             /**
-           * Initial data of .owl-stage
-           */
+             * Initial data of .owl-stage
+             */
             this.stageData = {
                 transform: 'translate3d(0px,0px,0px)',
                 transition: '0s',
@@ -431,44 +431,45 @@
              * All real items.
              */
             this._items = []; // is equal to this.slides
+            // is equal to this.slides
             /**
-           * Array with width of every slide.
-           */
+             * Array with width of every slide.
+             */
             this._widths = [];
             /**
-           * Currently suppressed events to prevent them from beeing retriggered.
-           */
+             * Currently suppressed events to prevent them from beeing retriggered.
+             */
             this._supress = {};
             /**
              * References to the running plugins of this carousel.
              */
             this._plugins = {};
             /**
-           * Absolute current position.
-           */
+             * Absolute current position.
+             */
             this._current = null;
             /**
-           * All cloned items.
-           */
+             * All cloned items.
+             */
             this._clones = [];
             /**
              * Merge values of all items.
-             * @todo Maybe this could be part of a plugin.
+             * \@todo Maybe this could be part of a plugin.
              */
             this._mergers = [];
             /**
-           * Animation speed in milliseconds.
-           */
+             * Animation speed in milliseconds.
+             */
             this._speed = null;
             /**
-           * Coordinates of all items in pixel.
-           * @todo The name of this member is missleading.
-           */
+             * Coordinates of all items in pixel.
+             * \@todo The name of this member is missleading.
+             */
             this._coordinates = [];
             /**
-           * Current breakpoint.
-           * @todo Real media queries would be nice.
-           */
+             * Current breakpoint.
+             * \@todo Real media queries would be nice.
+             */
             this._breakpoint = null;
             /**
              * Prefix for id of cloned slides
@@ -495,7 +496,7 @@
             };
             /**
              * Ordered list of workers for the update process.
-           */
+             */
             this._pipe = [
                 // {
                 //   filter: ['width', 'settings'],
@@ -985,7 +986,6 @@
          */
         /**
          * Checks whether user's option are set properly. Cheking is based on typings;
-         * @private
          * @param {?} options options set by user
          * @param {?} configOptions default options
          * @return {?} checked and modified (if it's needed) user's options
@@ -995,7 +995,6 @@
          */
         CarouselService.prototype._validateOptions = /**
          * Checks whether user's option are set properly. Cheking is based on typings;
-         * @private
          * @param {?} options options set by user
          * @param {?} configOptions default options
          * @return {?} checked and modified (if it's needed) user's options
@@ -1068,13 +1067,11 @@
          */
         /**
          * Checks option items set by user and if it bigger than number of slides then returns number of slides
-         * @private
          * @param {?} items option items set by user
          * @return {?} right number of items
          */
         CarouselService.prototype._validateItems = /**
          * Checks option items set by user and if it bigger than number of slides then returns number of slides
-         * @private
          * @param {?} items option items set by user
          * @return {?} right number of items
          */
@@ -1140,23 +1137,24 @@
                 this._defineSlidesData();
                 this.setOptions(options);
                 this.settings = __assign({}, this._options);
-                this.setViewportItemsN();
+                this.setOptionsForViewport();
                 this._trigger('change', { property: { name: 'settings', value: this.settings } });
                 this.invalidate('settings'); // must be call of this function;
                 this._trigger('changed', { property: { name: 'settings', value: this.settings } });
             };
         /**
-         * Set number of items for current viewport
+         * Set options for current viewport
          */
         /**
-         * Set number of items for current viewport
+         * Set options for current viewport
          * @return {?}
          */
-        CarouselService.prototype.setViewportItemsN = /**
-         * Set number of items for current viewport
+        CarouselService.prototype.setOptionsForViewport = /**
+         * Set options for current viewport
          * @return {?}
          */
             function () {
+                var _this = this;
                 /** @type {?} */
                 var viewport = this._width;
                 /** @type {?} */
@@ -1177,12 +1175,22 @@
                         }
                     }
                 }
-                this.settings = __assign({}, this.settings, { items: this._validateItems(overwrites[match].items) });
+                this.settings = __assign({}, this._options, overwrites[match], { items: (overwrites[match] && overwrites[match].items) ? this._validateItems(overwrites[match].items) : this._options.items });
                 // if (typeof this.settings.stagePadding === 'function') {
                 // 	this.settings.stagePadding = this.settings.stagePadding();
                 // }
                 delete this.settings.responsive;
                 this.owlDOMData.isResponsive = true;
+                this.owlDOMData.isMouseDragable = this.settings.mouseDrag;
+                this.owlDOMData.isTouchDragable = this.settings.touchDrag;
+                /** @type {?} */
+                var mergers = [];
+                this._items.forEach(function (item) {
+                    /** @type {?} */
+                    var mergeN = _this.settings.merge ? item.dataMerge : 1;
+                    mergers.push(mergeN);
+                });
+                this._mergers = mergers;
                 this._breakpoint = match;
                 this.invalidate('settings');
             };
@@ -1249,12 +1257,10 @@
            */
         /**
          * Updates option logic if necessery
-         * @private
          * @return {?}
          */
         CarouselService.prototype._optionsLogic = /**
          * Updates option logic if necessery
-         * @private
          * @return {?}
          */
             function () {
@@ -1339,7 +1345,7 @@
                 this.enter('refreshing');
                 this._trigger('refresh');
                 this._defineSlidesData();
-                this.setViewportItemsN();
+                this.setOptionsForViewport();
                 this._optionsLogic();
                 // this.$element.addClass(this.options.refreshClass);
                 this.update();
@@ -1568,7 +1574,7 @@
                 /** @type {?} */
                 var width = this.width();
                 /** @type {?} */
-                var coordinates = ((this.coordinates()));
+                var coordinates = ( /** @type {?} */(this.coordinates()));
                 /** @type {?} */
                 var position = -1;
                 if (this.settings.center) {
@@ -1999,7 +2005,7 @@
                 var result;
                 if (position === undefined) {
                     result = this._coordinates.map(function (item, index) {
-                        return ((_this.coordinates(index)));
+                        return ( /** @type {?} */(_this.coordinates(index)));
                     });
                     return result;
                 }
@@ -2026,7 +2032,6 @@
            */
         /**
          * Calculates the speed for a translation.
-         * @private
          * @param {?} from The absolute position of the start item.
          * @param {?} to The absolute position of the target item.
          * @param {?=} factor [factor=undefined] - The time factor in milliseconds.
@@ -2034,7 +2039,6 @@
          */
         CarouselService.prototype._duration = /**
          * Calculates the speed for a translation.
-         * @private
          * @param {?} from The absolute position of the start item.
          * @param {?} to The absolute position of the target item.
          * @param {?=} factor [factor=undefined] - The time factor in milliseconds.
@@ -2174,12 +2178,10 @@
            */
         /**
          * Gets viewport width.
-         * @private
          * @return {?} - The width in pixel.
          */
         CarouselService.prototype._viewport = /**
          * Gets viewport width.
-         * @private
          * @return {?} - The width in pixel.
          */
             function () {
@@ -2215,12 +2217,10 @@
          */
         /**
          * Sets slidesData using this._items
-         * @private
          * @return {?}
          */
         CarouselService.prototype._defineSlidesData = /**
          * Sets slidesData using this._items
-         * @private
          * @return {?}
          */
             function () {
@@ -2278,10 +2278,10 @@
                     'owl-animated-out': slide.isDefAnimatedOut
                 };
                 if (this.settings.animateIn) {
-                    currentClasses[((this.settings.animateIn))] = slide.isCustomAnimatedIn;
+                    currentClasses[( /** @type {?} */(this.settings.animateIn))] = slide.isCustomAnimatedIn;
                 }
                 if (this.settings.animateOut) {
-                    currentClasses[((this.settings.animateOut))] = slide.isCustomAnimatedOut;
+                    currentClasses[( /** @type {?} */(this.settings.animateOut))] = slide.isCustomAnimatedOut;
                 }
                 return currentClasses;
             };
@@ -2294,7 +2294,6 @@
            */
         /**
          * Operators to calculate right-to-left and left-to-right.
-         * @private
          * @param {?} a - The left side operand.
          * @param {?} o - The operator.
          * @param {?} b - The right side operand.
@@ -2302,7 +2301,6 @@
          */
         CarouselService.prototype._op = /**
          * Operators to calculate right-to-left and left-to-right.
-         * @private
          * @param {?} a - The left side operand.
          * @param {?} o - The operator.
          * @param {?} b - The right side operand.
@@ -2336,7 +2334,6 @@
         /**
          * Triggers a public event.
          * \@todo Remove `status`, `relatedTarget` should be used instead.
-         * @private
          * @param {?} name The event name.
          * @param {?=} data The event data.
          * @param {?=} namespace The event namespace.
@@ -2347,7 +2344,6 @@
         CarouselService.prototype._trigger = /**
          * Triggers a public event.
          * \@todo Remove `status`, `relatedTarget` should be used instead.
-         * @private
          * @param {?} name The event name.
          * @param {?=} data The event data.
          * @param {?=} namespace The event namespace.
@@ -2473,13 +2469,11 @@
            */
         /**
          * Suppresses events.
-         * @private
          * @param {?} events The events to suppress.
          * @return {?}
          */
         CarouselService.prototype._suppress = /**
          * Suppresses events.
-         * @private
          * @param {?} events The events to suppress.
          * @return {?}
          */
@@ -2495,13 +2489,11 @@
            */
         /**
          * Releases suppressed events.
-         * @private
          * @param {?} events The events to release.
          * @return {?}
          */
         CarouselService.prototype._release = /**
          * Releases suppressed events.
-         * @private
          * @param {?} events The events to release.
          * @return {?}
          */
@@ -2553,13 +2545,11 @@
            */
         /**
          * Determines if the input is a Number or something that can be coerced to a Number
-         * @private
          * @param {?} number The input to be tested
          * @return {?} An indication if the input is a Number or can be coerced to a Number
          */
         CarouselService.prototype._isNumeric = /**
          * Determines if the input is a Number or something that can be coerced to a Number
-         * @private
          * @param {?} number The input to be tested
          * @return {?} An indication if the input is a Number or can be coerced to a Number
          */
@@ -2573,13 +2563,11 @@
          */
         /**
          * Determines whether value is number or boolean type
-         * @private
          * @param {?} value The input to be tested
          * @return {?} An indication if the input is a Number or can be coerced to a Number, or Boolean
          */
         CarouselService.prototype._isNumberOrBoolean = /**
          * Determines whether value is number or boolean type
-         * @private
          * @param {?} value The input to be tested
          * @return {?} An indication if the input is a Number or can be coerced to a Number, or Boolean
          */
@@ -2593,13 +2581,11 @@
          */
         /**
          * Determines whether value is number or string type
-         * @private
          * @param {?} value The input to be tested
          * @return {?} An indication if the input is a Number or can be coerced to a Number, or String
          */
         CarouselService.prototype._isNumberOrString = /**
          * Determines whether value is number or string type
-         * @private
          * @param {?} value The input to be tested
          * @return {?} An indication if the input is a Number or can be coerced to a Number, or String
          */
@@ -2613,13 +2599,11 @@
          */
         /**
          * Determines whether value is number or string type
-         * @private
          * @param {?} value The input to be tested
          * @return {?} An indication if the input is a Number or can be coerced to a Number, or String
          */
         CarouselService.prototype._isStringOrBoolean = /**
          * Determines whether value is number or string type
-         * @private
          * @param {?} value The input to be tested
          * @return {?} An indication if the input is a Number or can be coerced to a Number, or String
          */
@@ -2656,6 +2640,7 @@
         CarouselService.decorators = [
             { type: core.Injectable }
         ];
+        /** @nocollapse */
         CarouselService.ctorParameters = function () {
             return [
                 { type: OwlLogger }
@@ -2666,9 +2651,9 @@
 
     /**
      * @fileoverview added by tsickle
-     * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+     * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
      */
-    var NavigationService = (function () {
+    var NavigationService = /** @class */ (function () {
         function NavigationService(carouselService) {
             this.carouselService = carouselService;
             /**
@@ -2779,12 +2764,10 @@
          */
         /**
          * Calculates internal states and updates prop _pages
-         * @private
          * @return {?}
          */
         NavigationService.prototype._updateNavPages = /**
          * Calculates internal states and updates prop _pages
-         * @private
          * @return {?}
          */
             function () {
@@ -2823,7 +2806,7 @@
                             }
                             j = 0, ++k;
                         }
-                        j += ((this.carouselService.mergers(this.carouselService.relative(i))));
+                        j += ( /** @type {?} */(this.carouselService.mergers(this.carouselService.relative(i))));
                     }
                 }
                 this._pages = pages;
@@ -2874,6 +2857,7 @@
                             this._dotsData.dots.push({
                                 active: false,
                                 id: "dot-" + (i + startI),
+                                innerContent: '',
                                 showInnerContent: false
                             });
                         }
@@ -2905,12 +2889,10 @@
          */
         /**
          * Changes state of nav buttons (disabled, enabled)
-         * @private
          * @return {?}
          */
         NavigationService.prototype._updateNavButtons = /**
          * Changes state of nav buttons (disabled, enabled)
-         * @private
          * @return {?}
          */
             function () {
@@ -2931,17 +2913,18 @@
          */
         /**
          * Changes active dot if page becomes changed
-         * @private
          * @return {?}
          */
         NavigationService.prototype._updateDots = /**
          * Changes active dot if page becomes changed
-         * @private
          * @return {?}
          */
             function () {
                 /** @type {?} */
                 var curActiveDotI;
+                if (!this.carouselService.settings.dots) {
+                    return;
+                }
                 this._dotsData.dots.forEach(function (item) {
                     if (item.active === true) {
                         item.active = false;
@@ -2959,12 +2942,10 @@
            */
         /**
          * Gets the current page position of the carousel.
-         * @private
          * @return {?} the current page position of the carousel
          */
         NavigationService.prototype._current = /**
          * Gets the current page position of the carousel.
-         * @private
          * @return {?} the current page position of the carousel
          */
             function () {
@@ -2988,13 +2969,11 @@
            */
         /**
          * Gets the current succesor/predecessor position.
-         * @private
          * @param {?} successor
          * @return {?} the current succesor/predecessor position
          */
         NavigationService.prototype._getPosition = /**
          * Gets the current succesor/predecessor position.
-         * @private
          * @param {?} successor
          * @return {?} the current succesor/predecessor position
          */
@@ -3126,6 +3105,7 @@
         NavigationService.decorators = [
             { type: core.Injectable }
         ];
+        /** @nocollapse */
         NavigationService.ctorParameters = function () {
             return [
                 { type: CarouselService }
@@ -3136,7 +3116,7 @@
 
     /**
      * @fileoverview added by tsickle
-     * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+     * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
      */
     /**
      * Create a new injection token for injecting the window into a component.
@@ -3150,10 +3130,7 @@
     var /**
      * Define abstract class for obtaining reference to the global window object.
      * @abstract
-     */ WindowRef = (function () {
-        /**
-         * Define abstract class for obtaining reference to the global window object.
-         */
+     */ WindowRef = /** @class */ (function () {
         function WindowRef() {
         }
         Object.defineProperty(WindowRef.prototype, "nativeWindow", {
@@ -3172,7 +3149,7 @@
      */
     var /**
      * Define class that implements the abstract class and returns the native window object.
-     */ BrowserWindowRef = (function (_super) {
+     */ BrowserWindowRef = /** @class */ (function (_super) {
         __extends(BrowserWindowRef, _super);
         function BrowserWindowRef() {
             return _super.call(this) || this;
@@ -3201,7 +3178,12 @@
         if (common.isPlatformBrowser(platformId)) {
             return browserWindowRef.nativeWindow;
         }
-        return new Object();
+        /** @type {?} */
+        var obj = {
+            setTimeout: function (func, time) { },
+            clearTimeout: function (a) { }
+        };
+        return obj;
     }
     /**
      * Create a injectable provider for the WindowRef token that uses the BrowserWindowRef class.
@@ -3228,7 +3210,7 @@
 
     /**
      * @fileoverview added by tsickle
-     * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+     * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
      */
     /**
      * Create a new injection token for injecting the Document into a component.
@@ -3242,10 +3224,7 @@
     var /**
      * Define abstract class for obtaining reference to the global Document object.
      * @abstract
-     */ DocumentRef = (function () {
-        /**
-         * Define abstract class for obtaining reference to the global Document object.
-         */
+     */ DocumentRef = /** @class */ (function () {
         function DocumentRef() {
         }
         Object.defineProperty(DocumentRef.prototype, "nativeDocument", {
@@ -3264,7 +3243,7 @@
      */
     var /**
      * Define class that implements the abstract class and returns the native Document object.
-     */ BrowserDocumentRef = (function (_super) {
+     */ BrowserDocumentRef = /** @class */ (function (_super) {
         __extends(BrowserDocumentRef, _super);
         function BrowserDocumentRef() {
             return _super.call(this) || this;
@@ -3293,7 +3272,12 @@
         if (common.isPlatformBrowser(platformId)) {
             return browserDocumentRef.nativeDocument;
         }
-        return new Object();
+        /** @type {?} */
+        var doc = {
+            hidden: false,
+            visibilityState: 'visible'
+        };
+        return doc;
     }
     /**
      * Create a injectable provider for the DocumentRef token that uses the BrowserDocumentRef class.
@@ -3320,9 +3304,9 @@
 
     /**
      * @fileoverview added by tsickle
-     * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+     * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
      */
-    var AutoplayService = (function () {
+    var AutoplayService = /** @class */ (function () {
         function AutoplayService(carouselService, winRef, docRef) {
             this.carouselService = carouselService;
             /**
@@ -3333,8 +3317,8 @@
              * Indicates whenever the autoplay is paused.
              */
             this._paused = false;
-            this.winRef = ((winRef));
-            this.docRef = ((docRef));
+            this.winRef = ( /** @type {?} */(winRef));
+            this.docRef = ( /** @type {?} */(docRef));
             this.spyDataStreams();
         }
         /**
@@ -3369,10 +3353,19 @@
                 var changedSettings$ = this.carouselService.getChangedState().pipe(operators.tap(function (data) {
                     _this._handleChangeObservable(data);
                 }));
+                /** @type {?} */
+                var resized$ = this.carouselService.getResizedState().pipe(operators.tap(function () {
+                    if (_this.carouselService.settings.autoplay) {
+                        _this.play();
+                    }
+                    else {
+                        _this.stop();
+                    }
+                }));
                 // original Autoplay Plugin has listeners on play.owl.core and stop.owl.core events.
                 // They are triggered by Video Plugin
                 /** @type {?} */
-                var autoplayMerge$ = rxjs.merge(initializedCarousel$, changedSettings$);
+                var autoplayMerge$ = rxjs.merge(initializedCarousel$, changedSettings$, resized$);
                 this.autoplaySubscription = autoplayMerge$.subscribe(function () { });
             };
         /**
@@ -3395,7 +3388,7 @@
             function (timeout, speed) {
                 if (this._paused) {
                     this._paused = false;
-                    this._setAutoPlayInterval();
+                    this._setAutoPlayInterval(1);
                 }
                 if (this.carouselService.is('rotating')) {
                     return;
@@ -3411,14 +3404,12 @@
            */
         /**
          * Gets a new timeout
-         * @private
          * @param {?=} timeout - The interval before the next animation starts.
          * @param {?=} speed - The animation speed for the animations.
          * @return {?}
          */
         AutoplayService.prototype._getNextTimeout = /**
          * Gets a new timeout
-         * @private
          * @param {?=} timeout - The interval before the next animation starts.
          * @param {?=} speed - The animation speed for the animations.
          * @return {?}
@@ -3428,6 +3419,7 @@
                 if (this._timeout) {
                     this.winRef.clearTimeout(this._timeout);
                 }
+                this._isArtificialAutoplayTimeout = timeout ? true : false;
                 return this.winRef.setTimeout(function () {
                     if (_this._paused || _this.carouselService.is('busy') || _this.carouselService.is('interacting') || _this.docRef.hidden) {
                         return;
@@ -3440,16 +3432,16 @@
            */
         /**
          * Sets autoplay in motion.
-         * @private
+         * @param {?=} timeout
          * @return {?}
          */
         AutoplayService.prototype._setAutoPlayInterval = /**
          * Sets autoplay in motion.
-         * @private
+         * @param {?=} timeout
          * @return {?}
          */
-            function () {
-                this._timeout = this._getNextTimeout();
+            function (timeout) {
+                this._timeout = this._getNextTimeout(timeout);
             };
         /**
          * Stops the autoplay.
@@ -3466,6 +3458,7 @@
                 if (!this.carouselService.is('rotating')) {
                     return;
                 }
+                this._paused = true;
                 this.winRef.clearTimeout(this._timeout);
                 this.carouselService.leave('rotating');
             };
@@ -3492,13 +3485,11 @@
          */
         /**
          * Manages by autoplaying according to data passed by _changedSettingsCarousel$ Obsarvable
-         * @private
          * @param {?} data object with current position of carousel and type of change
          * @return {?}
          */
         AutoplayService.prototype._handleChangeObservable = /**
          * Manages by autoplaying according to data passed by _changedSettingsCarousel$ Obsarvable
-         * @private
          * @param {?} data object with current position of carousel and type of change
          * @return {?}
          */
@@ -3517,6 +3508,21 @@
                         this._setAutoPlayInterval();
                     }
                 }
+            };
+        /**
+         * Starts autoplaying of the carousel in the case when user leaves the carousel before it starts translateing (moving)
+         */
+        /**
+         * Starts autoplaying of the carousel in the case when user leaves the carousel before it starts translateing (moving)
+         * @return {?}
+         */
+        AutoplayService.prototype._playAfterTranslated = /**
+         * Starts autoplaying of the carousel in the case when user leaves the carousel before it starts translateing (moving)
+         * @return {?}
+         */
+            function () {
+                var _this = this;
+                rxjs.of('translated').pipe(operators.switchMap(function (data) { return _this.carouselService.getTranslatedState(); }), operators.first(), operators.filter(function () { return _this._isArtificialAutoplayTimeout; }), operators.tap(function () { return _this._setAutoPlayInterval(); })).subscribe(function () { });
             };
         /**
          * Starts pausing
@@ -3547,7 +3553,8 @@
          */
             function () {
                 if (this.carouselService.settings.autoplayHoverPause && this.carouselService.is('rotating')) {
-                    this.pause();
+                    this.play();
+                    this._playAfterTranslated();
                 }
             };
         /**
@@ -3563,12 +3570,14 @@
          */
             function () {
                 if (this.carouselService.settings.autoplayHoverPause && this.carouselService.is('rotating')) {
-                    this.pause();
+                    this.play();
+                    this._playAfterTranslated();
                 }
             };
         AutoplayService.decorators = [
             { type: core.Injectable }
         ];
+        /** @nocollapse */
         AutoplayService.ctorParameters = function () {
             return [
                 { type: CarouselService },
@@ -3581,9 +3590,9 @@
 
     /**
      * @fileoverview added by tsickle
-     * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+     * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
      */
-    var LazyLoadService = (function () {
+    var LazyLoadService = /** @class */ (function () {
         function LazyLoadService(carouselService) {
             this.carouselService = carouselService;
             this.spyDataStreams();
@@ -3625,12 +3634,10 @@
                 this.lazyLoadSubscription = lazyLoadMerge$.subscribe(function () { });
             };
         /**
-         * @private
          * @param {?} data
          * @return {?}
          */
         LazyLoadService.prototype._defineLazyLoadSlides = /**
-         * @private
          * @param {?} data
          * @return {?}
          */
@@ -3675,13 +3682,11 @@
            */
         /**
          * Loads all resources of an item at the specified position.
-         * @private
          * @param {?} position - The absolute position of the item.
          * @return {?}
          */
         LazyLoadService.prototype._load = /**
          * Loads all resources of an item at the specified position.
-         * @private
          * @param {?} position - The absolute position of the item.
          * @return {?}
          */
@@ -3694,6 +3699,7 @@
         LazyLoadService.decorators = [
             { type: core.Injectable }
         ];
+        /** @nocollapse */
         LazyLoadService.ctorParameters = function () {
             return [
                 { type: CarouselService }
@@ -3704,9 +3710,9 @@
 
     /**
      * @fileoverview added by tsickle
-     * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+     * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
      */
-    var AnimateService = (function () {
+    var AnimateService = /** @class */ (function () {
         function AnimateService(carouselService) {
             this.carouselService = carouselService;
             /**
@@ -3776,12 +3782,10 @@
            */
         /**
          * Toggles the animation classes whenever an translations starts.
-         * @private
          * @return {?}
          */
         AnimateService.prototype._swap = /**
          * Toggles the animation classes whenever an translations starts.
-         * @private
          * @return {?}
          */
             function () {
@@ -3858,6 +3862,7 @@
         AnimateService.decorators = [
             { type: core.Injectable }
         ];
+        /** @nocollapse */
         AnimateService.ctorParameters = function () {
             return [
                 { type: CarouselService }
@@ -3868,9 +3873,9 @@
 
     /**
      * @fileoverview added by tsickle
-     * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+     * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
      */
-    var AutoHeightService = (function () {
+    var AutoHeightService = /** @class */ (function () {
         function AutoHeightService(carouselService) {
             this.carouselService = carouselService;
             this.spyDataStreams();
@@ -3951,6 +3956,7 @@
         AutoHeightService.decorators = [
             { type: core.Injectable }
         ];
+        /** @nocollapse */
         AutoHeightService.ctorParameters = function () {
             return [
                 { type: CarouselService }
@@ -3961,14 +3967,24 @@
 
     /**
      * @fileoverview added by tsickle
-     * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+     * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
      */
-    var HashService = (function () {
+    var HashService = /** @class */ (function () {
         function HashService(carouselService, route, router$$1) {
             this.carouselService = carouselService;
             this.route = route;
             this.router = router$$1;
             this.spyDataStreams();
+            if (!this.route) {
+                this.route = ( /** @type {?} */({
+                    fragment: rxjs.of('no route').pipe(operators.take(1))
+                }));
+            }
+            if (!this.router) {
+                this.router = ( /** @type {?} */({
+                    navigate: function (commands, extras) { return; }
+                }));
+            }
         }
         /**
          * @return {?}
@@ -4057,11 +4073,12 @@
         HashService.decorators = [
             { type: core.Injectable }
         ];
+        /** @nocollapse */
         HashService.ctorParameters = function () {
             return [
                 { type: CarouselService },
-                { type: router.ActivatedRoute },
-                { type: router.Router }
+                { type: router.ActivatedRoute, decorators: [{ type: core.Optional }] },
+                { type: router.Router, decorators: [{ type: core.Optional }] }
             ];
         };
         return HashService;
@@ -4069,11 +4086,11 @@
 
     /**
      * @fileoverview added by tsickle
-     * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+     * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
      */
     /** @type {?} */
     var nextId = 0;
-    var CarouselSlideDirective = (function () {
+    var CarouselSlideDirective = /** @class */ (function () {
         function CarouselSlideDirective(tplRef) {
             this.tplRef = tplRef;
             /**
@@ -4133,6 +4150,7 @@
         CarouselSlideDirective.decorators = [
             { type: core.Directive, args: [{ selector: 'ng-template[carouselSlide]' },] }
         ];
+        /** @nocollapse */
         CarouselSlideDirective.ctorParameters = function () {
             return [
                 { type: core.TemplateRef }
@@ -4152,16 +4170,13 @@
      */
     var /**
      * Data which will be passed out after ending of transition of carousel
-     */ SlidesOutputData = (function () {
-        /**
-         * Data which will be passed out after ending of transition of carousel
-         */
+     */ SlidesOutputData = /** @class */ (function () {
         function SlidesOutputData() {
         }
         return SlidesOutputData;
     }());
-    var CarouselComponent = (function () {
-        function CarouselComponent(el, resizeService, carouselService, navigationService, autoplayService, lazyLoadService, animateService, autoHeightService, hashService, logger) {
+    var CarouselComponent = /** @class */ (function () {
+        function CarouselComponent(el, resizeService, carouselService, navigationService, autoplayService, lazyLoadService, animateService, autoHeightService, hashService, logger, docRef) {
             this.el = el;
             this.resizeService = resizeService;
             this.carouselService = carouselService;
@@ -4175,6 +4190,7 @@
             this.translated = new core.EventEmitter();
             this.dragging = new core.EventEmitter();
             this.change = new core.EventEmitter();
+            this.changed = new core.EventEmitter();
             this.initialized = new core.EventEmitter();
             /**
              *  Data of every slide
@@ -4184,7 +4200,30 @@
              * Shows whether carousel is loaded of not.
              */
             this.carouselLoaded = false;
+            this.docRef = ( /** @type {?} */(docRef));
         }
+        /**
+         * @param {?} ev
+         * @return {?}
+         */
+        CarouselComponent.prototype.onVisibilityChange = /**
+         * @param {?} ev
+         * @return {?}
+         */
+            function (ev) {
+                if (!this.carouselService.settings.autoplay)
+                    return;
+                switch (this.docRef.visibilityState) {
+                    case 'visible':
+                        this.autoplayService.play();
+                        break;
+                    case 'hidden':
+                        this.autoplayService.stop();
+                        break;
+                    default:
+                        break;
+                }
+            };
         /**
          * @return {?}
          */
@@ -4289,15 +4328,50 @@
                     _this.change.emit(_this.slidesOutputData);
                     // this.slidesOutputData = {};
                 }));
+                this._changedCarousel$ = this.carouselService.getChangeState().pipe(operators.switchMap(function (value) {
+                    /** @type {?} */
+                    var changedPosition = rxjs.of(value).pipe(operators.filter(function () { return value.property.name === 'position'; }), operators.switchMap(function () { return rxjs.from(_this.slidesData); }), operators.skip(value.property.value), operators.take(_this.carouselService.settings.items), operators.map(function (slide) {
+                        /** @type {?} */
+                        var clonedIdPrefix = _this.carouselService.clonedIdPrefix;
+                        /** @type {?} */
+                        var id = slide.id.indexOf(clonedIdPrefix) >= 0 ? slide.id.slice(clonedIdPrefix.length) : slide.id;
+                        return __assign({}, slide, { id: id, isActive: true });
+                    }), operators.toArray(), operators.map(function (slides) {
+                        return {
+                            slides: slides,
+                            startPosition: _this.carouselService.relative(value.property.value)
+                        };
+                    }));
+                    // const changedSetting: Observable<SlidesOutputData> = of(value).pipe(
+                    //   filter(() => value.property.name === 'settings'),
+                    //   map(() => {
+                    //     return {
+                    //       slides: [],
+                    //       startPosition: this.carouselService.relative(value.property.value)
+                    //     }
+                    //   })
+                    // )
+                    return rxjs.merge(changedPosition);
+                }), operators.tap(function (slidesData) {
+                    _this.gatherTranslatedData();
+                    _this.changed.emit(slidesData.slides.length ? slidesData : _this.slidesOutputData);
+                    // console.log(this.slidesOutputData);
+                    // this.slidesOutputData = {};
+                }));
                 this._draggingCarousel$ = this.carouselService.getDragState().pipe(operators.tap(function () {
                     _this.gatherTranslatedData();
                     _this.dragging.emit({ dragging: true, data: _this.slidesOutputData });
-                }), operators.switchMap(function () {
-                    return _this.carouselService.getTranslatedState().pipe(operators.first(), operators.tap(function () {
-                        _this.dragging.emit({ dragging: false, data: _this.slidesOutputData });
-                    }));
+                }), operators.switchMap(function () { return _this.carouselService.getDraggedState().pipe(operators.map(function () { return !!_this.carouselService.is('animating'); })); }), operators.switchMap(function (anim) {
+                    if (anim) {
+                        return _this.carouselService.getTranslatedState().pipe(operators.first());
+                    }
+                    else {
+                        return rxjs.of('not animating');
+                    }
+                }), operators.tap(function () {
+                    _this.dragging.emit({ dragging: false, data: _this.slidesOutputData });
                 }));
-                this._carouselMerge$ = rxjs.merge(this._viewCurSettings$, this._translatedCarousel$, this._draggingCarousel$, this._changeCarousel$, this._initializedCarousel$);
+                this._carouselMerge$ = rxjs.merge(this._viewCurSettings$, this._translatedCarousel$, this._draggingCarousel$, this._changeCarousel$, this._changedCarousel$, this._initializedCarousel$);
                 this._allObservSubscription = this._carouselMerge$.subscribe(function () { });
             };
         /**
@@ -4305,12 +4379,10 @@
          */
         /**
          * Init subscription to resize event and attaches handler for this event
-         * @private
          * @return {?}
          */
         CarouselComponent.prototype._winResizeWatcher = /**
          * Init subscription to resize event and attaches handler for this event
-         * @private
          * @return {?}
          */
             function () {
@@ -4350,7 +4422,7 @@
          * @return {?}
          */
             function () {
-                if (!this.carouselLoaded || (this.navData && this.navData.disabled))
+                if (!this.carouselLoaded)
                     return;
                 this.navigationService.next(this.carouselService.settings.navSpeed);
             };
@@ -4366,7 +4438,7 @@
          * @return {?}
          */
             function () {
-                if (!this.carouselLoaded || (this.navData && this.navData.disabled))
+                if (!this.carouselLoaded)
                     return;
                 this.navigationService.prev(this.carouselService.settings.navSpeed);
             };
@@ -4403,7 +4475,8 @@
          * @return {?}
          */
             function (id) {
-                if (!this.carouselLoaded || (this.navData && this.navData.disabled) || (this.dotsData && this.dotsData.disabled))
+                // if (!this.carouselLoaded || ((this.navData && this.navData.disabled) && (this.dotsData && this.dotsData.disabled))) return;
+                if (!this.carouselLoaded)
                     return;
                 this.navigationService.toSlideById(id);
             };
@@ -4501,6 +4574,7 @@
                         styles: [".owl-theme { display: block; }"]
                     }] }
         ];
+        /** @nocollapse */
         CarouselComponent.ctorParameters = function () {
             return [
                 { type: core.ElementRef },
@@ -4512,7 +4586,8 @@
                 { type: AnimateService },
                 { type: AutoHeightService },
                 { type: HashService },
-                { type: OwlLogger }
+                { type: OwlLogger },
+                { type: undefined, decorators: [{ type: core.Inject, args: [DOCUMENT,] }] }
             ];
         };
         CarouselComponent.propDecorators = {
@@ -4520,17 +4595,19 @@
             translated: [{ type: core.Output }],
             dragging: [{ type: core.Output }],
             change: [{ type: core.Output }],
+            changed: [{ type: core.Output }],
             initialized: [{ type: core.Output }],
-            options: [{ type: core.Input }]
+            options: [{ type: core.Input }],
+            onVisibilityChange: [{ type: core.HostListener, args: ['document:visibilitychange', ['$event'],] }]
         };
         return CarouselComponent;
     }());
 
     /**
      * @fileoverview added by tsickle
-     * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+     * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
      */
-    var StageComponent = (function () {
+    var StageComponent = /** @class */ (function () {
         function StageComponent(zone, el, renderer, carouselService, animateService) {
             var _this = this;
             this.zone = zone;
@@ -4676,14 +4753,12 @@
         /**
          * Handles `touchstart` and `mousedown` events.
          * \@todo Horizontal swipe threshold as option / #261
-         * @private
          * @param {?} event - The event arguments.
          * @return {?}
          */
         StageComponent.prototype._onDragStart = /**
          * Handles `touchstart` and `mousedown` events.
          * \@todo Horizontal swipe threshold as option / #261
-         * @private
          * @param {?} event - The event arguments.
          * @return {?}
          */
@@ -4700,7 +4775,6 @@
                 this._drag.stage.start = stage;
                 this._drag.stage.current = stage;
                 this._drag.pointer = this._pointer(event);
-                this._drag.active = true;
                 this.listenerMouseUp = this.renderer.listen(document, 'mouseup', this.bindOnDragEnd);
                 this.listenerTouchEnd = this.renderer.listen(document, 'touchend', this.bindOnDragEnd);
                 this.zone.runOutsideAngular(function () {
@@ -4714,30 +4788,28 @@
          */
         /**
          * Attaches listeners to `touchmove` and `mousemove` events; initiates updating carousel after starting dragging
-         * @private
          * @param {?} event event objech of mouse or touch event
          * @return {?}
          */
         StageComponent.prototype._oneMouseTouchMove = /**
          * Attaches listeners to `touchmove` and `mousemove` events; initiates updating carousel after starting dragging
-         * @private
          * @param {?} event event objech of mouse or touch event
          * @return {?}
          */
             function (event) {
-                if (!this._drag.active)
-                    return false;
                 /** @type {?} */
                 var delta = this._difference(this._drag.pointer, this._pointer(event));
                 if (this.listenerATag) {
                     this.listenerATag();
                 }
-                this.listenerOneMouseMove();
-                this.listenerOneTouchMove();
-                if (Math.abs(delta.x) < Math.abs(delta.y) && this._is('valid')) {
-                    this._drag.active = false;
+                if (Math.abs(delta.x) < 3 && Math.abs(delta.y) < 3 && this._is('valid')) {
                     return;
                 }
+                if ((Math.abs(delta.x) < 3 && Math.abs(delta.x) < Math.abs(delta.y)) && this._is('valid')) {
+                    return;
+                }
+                this.listenerOneMouseMove();
+                this.listenerOneTouchMove();
                 this._drag.moving = true;
                 this.blockClickAnchorInDragging(event);
                 this.listenerMouseMove = this.renderer.listen(document, 'mousemove', this.bindOnDragMove);
@@ -4753,13 +4825,11 @@
          */
         /**
          * Attaches handler to HTMLAnchorElement for preventing click while carousel is being dragged
-         * @private
          * @param {?} event event object
          * @return {?}
          */
         StageComponent.prototype.blockClickAnchorInDragging = /**
          * Attaches handler to HTMLAnchorElement for preventing click while carousel is being dragged
-         * @private
          * @param {?} event event object
          * @return {?}
          */
@@ -4781,20 +4851,16 @@
         /**
          * Handles the `touchmove` and `mousemove` events.
          * \@todo #261
-         * @private
          * @param {?} event - The event arguments.
          * @return {?}
          */
         StageComponent.prototype._onDragMove = /**
          * Handles the `touchmove` and `mousemove` events.
          * \@todo #261
-         * @private
          * @param {?} event - The event arguments.
          * @return {?}
          */
             function (event) {
-                if (!this._drag.active)
-                    return false;
                 /** @type {?} */
                 var stage;
                 /** @type {?} */
@@ -4802,7 +4868,7 @@
                 if (stageOrExit === false) {
                     return;
                 }
-                stage = ((stageOrExit));
+                stage = ( /** @type {?} */(stageOrExit));
                 event.preventDefault();
                 this._drag.stage.current = stage;
                 this._animate(stage.x - this._drag.stage.start.x);
@@ -4813,13 +4879,11 @@
          */
         /**
          * Moves .owl-stage left-right
-         * @private
          * @param {?} coordinate coordinate to be set to .owl-stage
          * @return {?}
          */
         StageComponent.prototype._animate = /**
          * Moves .owl-stage left-right
-         * @private
          * @param {?} coordinate coordinate to be set to .owl-stage
          * @return {?}
          */
@@ -4836,19 +4900,19 @@
         /**
          * Handles the `touchend` and `mouseup` events.
          * \@todo #261 / Threshold for click event
-         * @private
          * @param {?} event - The event arguments.
          * @return {?}
          */
         StageComponent.prototype._onDragEnd = /**
          * Handles the `touchend` and `mouseup` events.
          * \@todo #261 / Threshold for click event
-         * @private
          * @param {?} event - The event arguments.
          * @return {?}
          */
             function (event) {
                 this.carouselService.owlDOMData.isGrab = false;
+                this.listenerOneMouseMove();
+                this.listenerOneTouchMove();
                 if (this._drag.moving) {
                     this.renderer.setStyle(this.el.nativeElement.children[0], 'transform', "");
                     this.renderer.setStyle(this.el.nativeElement.children[0], 'transition', this.carouselService.speed(+this.carouselService.settings.dragEndSpeed || this.carouselService.settings.smartSpeed) / 1000 + 's');
@@ -4879,13 +4943,11 @@
            */
         /**
          * Prepares data for dragging carousel. It starts after firing `touchstart` and `mousedown` events.
-         * @private
          * @param {?} event - The event arguments.
          * @return {?} stage - object with 'x' and 'y' coordinates of .owl-stage
          */
         StageComponent.prototype._prepareDragging = /**
          * Prepares data for dragging carousel. It starts after firing `touchstart` and `mousedown` events.
-         * @private
          * @param {?} event - The event arguments.
          * @return {?} stage - object with 'x' and 'y' coordinates of .owl-stage
          */
@@ -4898,13 +4960,11 @@
          */
         /**
          * Finishes dragging
-         * @private
          * @param {?} event object event of 'mouseUp' of 'touchend' events
          * @return {?}
          */
         StageComponent.prototype._finishDragging = /**
          * Finishes dragging
-         * @private
          * @param {?} event object event of 'mouseUp' of 'touchend' events
          * @return {?}
          */
@@ -4918,13 +4978,11 @@
            */
         /**
          * Gets unified pointer coordinates from event.
-         * @private
          * @param {?} event The `mousedown` or `touchstart` event.
          * @return {?} Contains `x` and `y` coordinates of current pointer position.
          */
         StageComponent.prototype._pointer = /**
          * Gets unified pointer coordinates from event.
-         * @private
          * @param {?} event The `mousedown` or `touchstart` event.
          * @return {?} Contains `x` and `y` coordinates of current pointer position.
          */
@@ -4939,14 +4997,12 @@
            */
         /**
          * Gets the difference of two vectors.
-         * @private
          * @param {?} firstC
          * @param {?} second
          * @return {?} The difference.
          */
         StageComponent.prototype._difference = /**
          * Gets the difference of two vectors.
-         * @private
          * @param {?} firstC
          * @param {?} second
          * @return {?} The difference.
@@ -4961,13 +5017,11 @@
            */
         /**
          * Checks whether the carousel is in a specific state or not.
-         * @private
          * @param {?} specificState The state to check.
          * @return {?} The flag which indicates if the carousel is busy.
          */
         StageComponent.prototype._is = /**
          * Checks whether the carousel is in a specific state or not.
-         * @private
          * @param {?} specificState The state to check.
          * @return {?} The flag which indicates if the carousel is busy.
          */
@@ -4980,13 +5034,11 @@
         */
         /**
          * Enters a state.
-         * @private
          * @param {?} name The state name.
          * @return {?}
          */
         StageComponent.prototype._enter = /**
          * Enters a state.
-         * @private
          * @param {?} name The state name.
          * @return {?}
          */
@@ -4998,12 +5050,10 @@
            */
         /**
          * Sends all data needed for View.
-         * @private
          * @return {?}
          */
         StageComponent.prototype._sendChanges = /**
          * Sends all data needed for View.
-         * @private
          * @return {?}
          */
             function () {
@@ -5028,12 +5078,10 @@
            */
         /**
          * Enters into a 'dragging' state
-         * @private
          * @return {?}
          */
         StageComponent.prototype._enterDragging = /**
          * Enters into a 'dragging' state
-         * @private
          * @return {?}
          */
             function () {
@@ -5059,7 +5107,7 @@
         StageComponent.decorators = [
             { type: core.Component, args: [{
                         selector: 'owl-stage',
-                        template: "\n    <div>\n      <div class=\"owl-stage\" [ngStyle]=\"{'width': stageData.width + 'px',\n                                        'transform': stageData.transform,\n                                        'transition': stageData.transition,\n                                        'padding-left': stageData.paddingL + 'px',\n                                        'padding-right': stageData.paddingR + 'px' }\"\n          (transitionend)=\"onTransitionEnd()\">\n        <ng-container *ngFor=\"let slide of slidesData; let i = index\">\n          <div class=\"owl-item\" [ngClass]=\"slide.classes\"\n                                [ngStyle]=\"{'width': slide.width + 'px',\n                                            'margin-left': slide.marginL + 'px',\n                                            'margin-right': slide.marginR + 'px',\n                                            'left': slide.left}\"\n                                (animationend)=\"clear(slide.id)\"\n                                [@autoHeight]=\"slide.heightState\">\n            <ng-template *ngIf=\"slide.load\" [ngTemplateOutlet]=\"slide.tplRef\"></ng-template>\n          </div><!-- /.owl-item -->\n        </ng-container>\n      </div><!-- /.owl-stage -->\n    </div>\n  ",
+                        template: "\n    <div>\n      <div class=\"owl-stage\" [ngStyle]=\"{'width': stageData.width + 'px',\n                                        'transform': stageData.transform,\n                                        'transition': stageData.transition,\n                                        'padding-left': stageData.paddingL ? stageData.paddingL + 'px' : '',\n                                        'padding-right': stageData.paddingR ? stageData.paddingR + 'px' : '' }\"\n          (transitionend)=\"onTransitionEnd()\">\n        <ng-container *ngFor=\"let slide of slidesData; let i = index\">\n          <div class=\"owl-item\" [ngClass]=\"slide.classes\"\n                                [ngStyle]=\"{'width': slide.width + 'px',\n                                            'margin-left': slide.marginL ? slide.marginL + 'px' : '',\n                                            'margin-right': slide.marginR ? slide.marginR + 'px' : '',\n                                            'left': slide.left}\"\n                                (animationend)=\"clear(slide.id)\"\n                                [@autoHeight]=\"slide.heightState\">\n            <ng-template *ngIf=\"slide.load\" [ngTemplateOutlet]=\"slide.tplRef\"></ng-template>\n          </div><!-- /.owl-item -->\n        </ng-container>\n      </div><!-- /.owl-stage -->\n    </div>\n  ",
                         animations: [
                             animations.trigger('autoHeight', [
                                 animations.state('nulled', animations.style({ height: 0 })),
@@ -5076,6 +5124,7 @@
                         ]
                     }] }
         ];
+        /** @nocollapse */
         StageComponent.ctorParameters = function () {
             return [
                 { type: core.NgZone },
@@ -5100,9 +5149,9 @@
 
     /**
      * @fileoverview added by tsickle
-     * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+     * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
      */
-    var OwlRouterLinkDirective = (function () {
+    var OwlRouterLinkDirective = /** @class */ (function () {
         function OwlRouterLinkDirective(router$$1, route, tabIndex, renderer, el) {
             this.router = router$$1;
             this.route = route;
@@ -5136,7 +5185,7 @@
              * @param {?} value
              * @return {?}
              */ function (value) {
-                if (core.isDevMode() && ((console)) && ((console.warn))) {
+                if (core.isDevMode() && ( /** @type {?} */(console)) && ( /** @type {?} */(console.warn))) {
                     console.warn('preserveQueryParams is deprecated!, use queryParamsHandling instead.');
                 }
                 this.preserve = value;
@@ -5181,6 +5230,7 @@
         OwlRouterLinkDirective.decorators = [
             { type: core.Directive, args: [{ selector: ':not(a)[owlRouterLink]' },] }
         ];
+        /** @nocollapse */
         OwlRouterLinkDirective.ctorParameters = function () {
             return [
                 { type: router.Router },
@@ -5215,7 +5265,7 @@
      *
      * \@publicApi
      */
-    var OwlRouterLinkWithHrefDirective = (function () {
+    var OwlRouterLinkWithHrefDirective = /** @class */ (function () {
         function OwlRouterLinkWithHrefDirective(router$$1, route, locationStrategy) {
             var _this = this;
             this.router = router$$1;
@@ -5249,7 +5299,7 @@
              * @param {?} value
              * @return {?}
              */ function (value) {
-                if (core.isDevMode() && ((console)) && ((console.warn))) {
+                if (core.isDevMode() && ( /** @type {?} */(console)) && ( /** @type {?} */(console.warn))) {
                     console.warn('preserveQueryParams is deprecated, use queryParamsHandling instead.');
                 }
                 this.preserve = value;
@@ -5306,11 +5356,9 @@
                 return false;
             };
         /**
-         * @private
          * @return {?}
          */
         OwlRouterLinkWithHrefDirective.prototype.updateTargetUrlAndHref = /**
-         * @private
          * @return {?}
          */
             function () {
@@ -5335,6 +5383,7 @@
         OwlRouterLinkWithHrefDirective.decorators = [
             { type: core.Directive, args: [{ selector: 'a[owlRouterLink]' },] }
         ];
+        /** @nocollapse */
         OwlRouterLinkWithHrefDirective.ctorParameters = function () {
             return [
                 { type: router.Router },
@@ -5368,19 +5417,15 @@
 
     /**
      * @fileoverview added by tsickle
-     * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+     * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
      */
-    /** @type {?} */
-    var routes = [];
-    var CarouselModule = (function () {
+    var CarouselModule = /** @class */ (function () {
         function CarouselModule() {
         }
         CarouselModule.decorators = [
             { type: core.NgModule, args: [{
                         imports: [
                             common.CommonModule,
-                            // BrowserAnimationsModule, // there's an issue with this import while using lazy loading of module consuming this library. I don't remove it because it could be needed during future enhancement of this lib.
-                            router.RouterModule.forChild(routes)
                         ],
                         declarations: [CarouselComponent, CarouselSlideDirective, StageComponent, OwlRouterLinkDirective, OwlRouterLinkWithHrefDirective],
                         exports: [CarouselComponent, CarouselSlideDirective, OwlRouterLinkDirective, OwlRouterLinkWithHrefDirective],
@@ -5392,17 +5437,17 @@
 
     /**
      * @fileoverview added by tsickle
-     * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+     * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
      */
 
     /**
      * @fileoverview added by tsickle
-     * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+     * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
      */
 
     /**
      * @fileoverview added by tsickle
-     * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+     * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
      */
 
     exports.CarouselModule = CarouselModule;
