@@ -1,6 +1,6 @@
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 export class ResponsiveSettings {
 }
@@ -199,4 +199,4 @@ if (false) {
      */
     OwlOptions.prototype.URLhashListener;
 }
-//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoib3dsLW9wdGlvbnMubW9kZWwuanMiLCJzb3VyY2VSb290Ijoibmc6Ly9uZ3gtb3dsLWNhcm91c2VsLW8vIiwic291cmNlcyI6WyJsaWIvbW9kZWxzL293bC1vcHRpb25zLm1vZGVsLnRzIl0sIm5hbWVzIjpbXSwibWFwcGluZ3MiOiI7Ozs7QUFBQSxNQUFNO0NBRUw7QUFDRCxNQUFNO0lBaUxKLGdCQUFnQixDQUFDO0NBQ2xCOzs7Ozs7SUE5S0MsMkJBQWU7Ozs7O0lBSWYsMEJBQWU7Ozs7O0lBSWYsNEJBQWlCOzs7OztJQUlqQiw0QkFBaUI7Ozs7O0lBS2pCLCtCQUFvQjs7Ozs7SUFJcEIsK0JBQW9COzs7OztJQUlwQiw4QkFBbUI7Ozs7O0lBSW5CLDhCQUFtQjs7Ozs7SUFLbkIsNEJBQWdCOzs7OztJQUloQixrQ0FBc0I7Ozs7O0lBS3RCLDJCQUFnQjs7Ozs7SUFJaEIsOEJBQW1COzs7OztJQUluQiwrQkFBb0I7Ozs7O0lBS3BCLG1DQUFnQzs7Ozs7SUFJaEMseUJBQWM7Ozs7O0lBS2QsZ0NBQW9COzs7OztJQUlwQixnQ0FBcUI7Ozs7O0lBSXJCLGtDQUFnQzs7Ozs7SUFLaEMsZ0NBQWdDOzs7OztJQUloQywyQ0FBK0I7Ozs7O0lBTS9CLHlCQUFjOzs7OztJQUlkLDZCQUFtQjs7Ozs7SUFJbkIsOEJBQTRCOzs7OztJQUk1Qiw2QkFBMEI7Ozs7O0lBSTFCLDBCQUFlOzs7OztJQUlmLDhCQUE0Qjs7Ozs7SUFJNUIsOEJBQW1COzs7OztJQUluQiwrQkFBNkI7Ozs7O0lBTzdCLDhCQUFtQjs7Ozs7SUFLbkIscUNBQXlCOzs7OztJQUt6Qix3Q0FBNkI7Ozs7O0lBSzdCLG1DQUFpQzs7Ozs7SUFNakMsOEJBQW1COzs7OztJQUtuQixtQ0FBdUI7Ozs7O0lBTXZCLGdDQUE4Qjs7Ozs7SUFLOUIsK0JBQTZCOzs7OztJQU03QixnQ0FBcUI7Ozs7O0lBTXJCLHFDQUEwQiIsInNvdXJjZXNDb250ZW50IjpbImV4cG9ydCBjbGFzcyBSZXNwb25zaXZlU2V0dGluZ3Mge1xyXG4gIFtrZXk6IG51bWJlcl06IHsgaXRlbXM6IG51bWJlcn1cclxufVxyXG5leHBvcnQgY2xhc3MgT3dsT3B0aW9ucyB7XHJcbiAgLyoqXHJcbiAgICogVGhlIG51bWJlciBvZiBpdGVtcyB5b3Ugd2FudCB0byBzZWUgb24gdGhlIHNjcmVlbi5cclxuICAgKi9cclxuICBpdGVtcz86IG51bWJlcjtcclxuICAvKipcclxuICAgKiBJbmZpbml0eSBsb29wLiBEdXBsaWNhdGUgbGFzdCBhbmQgZmlyc3QgaXRlbXMgdG8gZ2V0IGxvb3AgaWxsdXNpb24uXHJcbiAgICovXHJcbiAgbG9vcD86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogQ2VudGVyIGl0ZW0uIFdvcmtzIHdlbGwgd2l0aCBldmVuIGFuIG9kZCBudW1iZXIgb2YgaXRlbXMuXHJcbiAgICovXHJcbiAgY2VudGVyPzogYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBHbyBiYWNrd2FyZHMgd2hlbiB0aGUgYm91bmRhcnkgaGFzIHJlYWNoZWQuXHJcbiAgICovXHJcbiAgcmV3aW5kPzogYm9vbGVhbjtcclxuXHJcbiAgLyoqXHJcbiAgICogTW91c2UgZHJhZy5cclxuICAgKi9cclxuICBtb3VzZURyYWc/OiBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIFRvdWNoIGRyYWdcclxuICAgKi9cclxuICB0b3VjaERyYWc/OiBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIFN0YWdlIHB1bGwgdG8gZWRnZVxyXG4gICAqL1xyXG4gIHB1bGxEcmFnPzogYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBJdGVtIHB1bGwgdG8gZWRnZS5cclxuICAgKi9cclxuICBmcmVlRHJhZz86IGJvb2xlYW47XHJcblxyXG4gIC8qKlxyXG4gICAqIG1hcmdpbi1yaWdodChweCkgb24gaXRlbVxyXG4gICAqL1xyXG4gIG1hcmdpbj86IG51bWJlcjtcclxuICAvKipcclxuICAgKiBQYWRkaW5nIGxlZnQgYW5kIHJpZ2h0IG9uIHN0YWdlIChjYW4gc2VlIG5laWdoYm91cnMpXHJcbiAgICovXHJcbiAgc3RhZ2VQYWRkaW5nPzogbnVtYmVyO1xyXG5cclxuICAvKipcclxuICAgKiBNZXJnZSBpdGVtcy4gV29ya3Mgd2l0aCBASW5wdXQgb3B0aW9uICdkYXRhTWVyZ2UnIG9mIENhcm91c2VsU2xpZGVEaXJlY3RpdmUuXHJcbiAgICovXHJcbiAgbWVyZ2U/OiBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIEZpdCBtZXJnZWQgaXRlbXMgaWYgc2NyZWVuIGlzIHNtYWxsZXIgdGhhbiBpdGVtcyB2YWx1ZS5cclxuICAgKi9cclxuICBtZXJnZUZpdD86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogU2V0IG5vbiBncmlkIGNvbnRlbnQuICBXb3JrcyB3aXRoIEBJbnB1dCBvcHRpb24gJ3dpZHRoJyBvZiBDYXJvdXNlbFNsaWRlRGlyZWN0aXZlLlxyXG4gICAqL1xyXG4gIGF1dG9XaWR0aD86IGJvb2xlYW47XHJcblxyXG4gIC8qKlxyXG4gICAqIFN0YXJ0IHBvc2l0aW9uXHJcbiAgICovXHJcbiAgc3RhcnRQb3NpdGlvbj86IG51bWJlciB8IHN0cmluZztcclxuICAvKipcclxuICAgKiBDaGFuZ2UgZGlyZWN0aW9uIGZyb20gUmlnaHQgdG8gbGVmdFxyXG4gICAqL1xyXG4gIHJ0bD86IGJvb2xlYW47XHJcblxyXG4gIC8qKlxyXG4gICAqIFNwZWVkIENhbGN1bGF0ZSB3aGlsZSBkcmFnZ2luZ1xyXG4gICAqL1xyXG4gIHNtYXJ0U3BlZWQ/OiBudW1iZXI7XHJcbiAgLyoqXHJcbiAgICogU3BlZWQgQ2FsY3VsYXRlIHdoaWxlIGRyYWdnaW5nXHJcbiAgICovXHJcbiAgZmx1aWRTcGVlZD86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogRHJhZyBlbmQgc3BlZWRcclxuICAgKi9cclxuICBkcmFnRW5kU3BlZWQ/OiBudW1iZXIgfCBib29sZWFuO1xyXG5cclxuICAvKipcclxuICAgKiBPYmplY3QgY29udGFpbmluZyByZXNwb25zaXZlIG9wdGlvbnMuIENhbiBiZSBzZXQgdG8gZmFsc2UgdG8gcmVtb3ZlIHJlc3BvbnNpdmUgY2FwYWJpbGl0aWVzLlxyXG4gICAqL1xyXG4gIHJlc3BvbnNpdmU/OiBSZXNwb25zaXZlU2V0dGluZ3M7XHJcbiAgLyoqXHJcbiAgICogUmVzcG9uc2l2ZSByZWZyZXNoIHJhdGUuIFBlcmlvZCBvZiB0aW1lIGJldHdlZW4gZmlyaW5nICdyZXNpemUnIGV2ZW50IGFuZCByZWZyZXNoaW5nIGNhcm91c2VsLlxyXG4gICAqL1xyXG4gIHJlc3BvbnNpdmVSZWZyZXNoUmF0ZT86IG51bWJlcjtcclxuXHJcbiAgLy8gTmF2aWdhdGlvblxyXG4gIC8qKlxyXG4gICAqIFNob3cgbmV4dC9wcmV2IGJ1dHRvbnMuXHJcbiAgICovXHJcbiAgbmF2PzogYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBUZXh0IGluIG5leHQvcHJldiBidXR0b25zLiBIVE1MIGFsbG93ZWQuXHJcbiAgICovXHJcbiAgbmF2VGV4dD86IHN0cmluZ1tdO1xyXG4gIC8qKlxyXG4gICAqIE5hdmlnYXRpb24gc3BlZWRcclxuICAgKi9cclxuICBuYXZTcGVlZD86IG51bWJlciB8IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogTmF2aWdhdGlvbiBzbGlkZSBieSB4LiAncGFnZScgc3RyaW5nIGNhbiBiZSBzZXQgdG8gc2xpZGUgYnkgcGFnZS5cclxuICAgKi9cclxuICBzbGlkZUJ5PzogbnVtYmVyIHwgc3RyaW5nO1xyXG4gIC8qKlxyXG4gICAqIFNob3cgZG90cyBuYXZpZ2F0aW9uXHJcbiAgICovXHJcbiAgZG90cz86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogU2hvdyBkb3RzIGVhY2ggeCBpdGVtXHJcbiAgICovXHJcbiAgZG90c0VhY2g/OiBudW1iZXIgfCBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIFVzZWQgd2l0aCBASW5wdXQgb3B0aW9uICdkb3RDb250ZW50JyBvZiBDYXJvdXNlbFNsaWRlRGlyZWN0aXZlLlxyXG4gICAqL1xyXG4gIGRvdHNEYXRhPzogYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBQYWdpbmF0aW9uIHNwZWVkLlxyXG4gICAqL1xyXG4gIGRvdHNTcGVlZD86IG51bWJlciB8IGJvb2xlYW47XHJcblxyXG5cclxuICAvLyBBVVRPUExBWSAqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqXHJcbiAgLyoqXHJcbiAgICogQXV0b3BsYXkuXHJcbiAgICovXHJcbiAgYXV0b3BsYXk/OiBib29sZWFuO1xyXG5cclxuICAvKipcclxuICAgKiBBdXRvcGxheSBpbnRlcnZhbCB0aW1lb3V0XHJcbiAgICovXHJcbiAgYXV0b3BsYXlUaW1lb3V0PzogbnVtYmVyO1xyXG5cclxuICAvKipcclxuICAgKiBQYXVzZSBvbiBtb3VzZSBob3ZlclxyXG4gICAqL1xyXG4gIGF1dG9wbGF5SG92ZXJQYXVzZT86IGJvb2xlYW47XHJcblxyXG4gIC8qKlxyXG4gICAqIEF1dG9wbGF5IHNwZWVkXHJcbiAgICovXHJcbiAgYXV0b3BsYXlTcGVlZD86IG51bWJlciB8IGJvb2xlYW47XHJcblxyXG4gIC8vIExBWlkgTE9BRElORyAqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqXHJcbiAgLyoqXHJcbiAgICogTGF6eSBsb2FkaW5nIG9mIHNsaWRlcyB3aXRoIGltYWdlc1xyXG4gICAqL1xyXG4gIGxhenlMb2FkPzogYm9vbGVhbjtcclxuXHJcbiAgLyoqXHJcbiAgICogSW5kaWNhdGVzIGhvdyBtYW55IGl0ZW1zIChzbGlkZXMgd2l0aCBpbWFnZXMpIHRvIHRoZSByaWdodCAoYW5kIGxlZnQsIHdoZW4gbG9vcCBpcyB0cnVlKSB3aWxsIGJlIHByZS1sb2FkZWRcclxuICAgKi9cclxuICBsYXp5TG9hZEVhZ2VyPzogbnVtYmVyO1xyXG5cclxuICAvLyBBTklNQVRFICoqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKipcclxuICAvKipcclxuICAgKiBDbGFzcyBmb3IgQ1NTMyBhbmltYXRpb24gb3V0LlxyXG4gICAqL1xyXG4gIGFuaW1hdGVPdXQ/OiBzdHJpbmcgfCBib29sZWFuO1xyXG5cclxuICAvKipcclxuICAgKiBDbGFzcyBmb3IgQ1NTMyBhbmltYXRpb24gaW4uXHJcbiAgICovXHJcbiAgYW5pbWF0ZUluPzogc3RyaW5nIHwgYm9vbGVhbjtcclxuXHJcbiAgLy8gQVVUT0hFSUdIVCAqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqXHJcbiAgLyoqXHJcbiAgICogRW5hYmxlcyBjaGFuZ2luZyBjYXJvdXNlbCBhY2NvcmRpbmcgdG8gdGhlIGJpZ2dlc3QgaGVpZ2h0IG9mIGFzcyBhY3RpdmUgc2xpZGVzIHByb3ZpZGluZyBzbGlkZXMgaGF2ZSBkaWZmZXJlbnQgaGVpZ2h0XHJcbiAgICovXHJcbiAgYXV0b0hlaWdodD86IGJvb2xlYW47XHJcblxyXG4gIC8vIEhBU0ggKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKipcclxuICAvKipcclxuICAgKiBFbmFibGVzIGxpc3RlbmluZyB0byB1cmwgaGFzaCAoZnJhZ21lbnQpIGNoYW5nZXNcclxuICAgKi9cclxuICBVUkxoYXNoTGlzdGVuZXI/OiBib29sZWFuO1xyXG4gIGNvbnN0cnVjdG9yKCkgeyB9XHJcbn1cclxuIl19
+//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoib3dsLW9wdGlvbnMubW9kZWwuanMiLCJzb3VyY2VSb290Ijoibmc6Ly9uZ3gtb3dsLWNhcm91c2VsLW8vIiwic291cmNlcyI6WyJsaWIvbW9kZWxzL293bC1vcHRpb25zLm1vZGVsLnRzIl0sIm5hbWVzIjpbXSwibWFwcGluZ3MiOiI7Ozs7QUFBQSxNQUFNLE9BQU8sa0JBQWtCO0NBRTlCO0FBQ0QsTUFBTSxPQUFPLFVBQVU7SUFpTHJCLGdCQUFnQixDQUFDO0NBQ2xCOzs7Ozs7SUE5S0MsMkJBQWU7Ozs7O0lBSWYsMEJBQWU7Ozs7O0lBSWYsNEJBQWlCOzs7OztJQUlqQiw0QkFBaUI7Ozs7O0lBS2pCLCtCQUFvQjs7Ozs7SUFJcEIsK0JBQW9COzs7OztJQUlwQiw4QkFBbUI7Ozs7O0lBSW5CLDhCQUFtQjs7Ozs7SUFLbkIsNEJBQWdCOzs7OztJQUloQixrQ0FBc0I7Ozs7O0lBS3RCLDJCQUFnQjs7Ozs7SUFJaEIsOEJBQW1COzs7OztJQUluQiwrQkFBb0I7Ozs7O0lBS3BCLG1DQUFnQzs7Ozs7SUFJaEMseUJBQWM7Ozs7O0lBS2QsZ0NBQW9COzs7OztJQUlwQixnQ0FBcUI7Ozs7O0lBSXJCLGtDQUFnQzs7Ozs7SUFLaEMsZ0NBQWdDOzs7OztJQUloQywyQ0FBK0I7Ozs7O0lBTS9CLHlCQUFjOzs7OztJQUlkLDZCQUFtQjs7Ozs7SUFJbkIsOEJBQTRCOzs7OztJQUk1Qiw2QkFBMEI7Ozs7O0lBSTFCLDBCQUFlOzs7OztJQUlmLDhCQUE0Qjs7Ozs7SUFJNUIsOEJBQW1COzs7OztJQUluQiwrQkFBNkI7Ozs7O0lBTzdCLDhCQUFtQjs7Ozs7SUFLbkIscUNBQXlCOzs7OztJQUt6Qix3Q0FBNkI7Ozs7O0lBSzdCLG1DQUFpQzs7Ozs7SUFNakMsOEJBQW1COzs7OztJQUtuQixtQ0FBdUI7Ozs7O0lBTXZCLGdDQUE4Qjs7Ozs7SUFLOUIsK0JBQTZCOzs7OztJQU03QixnQ0FBcUI7Ozs7O0lBTXJCLHFDQUEwQiIsInNvdXJjZXNDb250ZW50IjpbImV4cG9ydCBjbGFzcyBSZXNwb25zaXZlU2V0dGluZ3Mge1xyXG4gIFtrZXk6IG51bWJlcl06IHsgaXRlbXM6IG51bWJlcn1cclxufVxyXG5leHBvcnQgY2xhc3MgT3dsT3B0aW9ucyB7XHJcbiAgLyoqXHJcbiAgICogVGhlIG51bWJlciBvZiBpdGVtcyB5b3Ugd2FudCB0byBzZWUgb24gdGhlIHNjcmVlbi5cclxuICAgKi9cclxuICBpdGVtcz86IG51bWJlcjtcclxuICAvKipcclxuICAgKiBJbmZpbml0eSBsb29wLiBEdXBsaWNhdGUgbGFzdCBhbmQgZmlyc3QgaXRlbXMgdG8gZ2V0IGxvb3AgaWxsdXNpb24uXHJcbiAgICovXHJcbiAgbG9vcD86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogQ2VudGVyIGl0ZW0uIFdvcmtzIHdlbGwgd2l0aCBldmVuIGFuIG9kZCBudW1iZXIgb2YgaXRlbXMuXHJcbiAgICovXHJcbiAgY2VudGVyPzogYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBHbyBiYWNrd2FyZHMgd2hlbiB0aGUgYm91bmRhcnkgaGFzIHJlYWNoZWQuXHJcbiAgICovXHJcbiAgcmV3aW5kPzogYm9vbGVhbjtcclxuXHJcbiAgLyoqXHJcbiAgICogTW91c2UgZHJhZy5cclxuICAgKi9cclxuICBtb3VzZURyYWc/OiBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIFRvdWNoIGRyYWdcclxuICAgKi9cclxuICB0b3VjaERyYWc/OiBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIFN0YWdlIHB1bGwgdG8gZWRnZVxyXG4gICAqL1xyXG4gIHB1bGxEcmFnPzogYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBJdGVtIHB1bGwgdG8gZWRnZS5cclxuICAgKi9cclxuICBmcmVlRHJhZz86IGJvb2xlYW47XHJcblxyXG4gIC8qKlxyXG4gICAqIG1hcmdpbi1yaWdodChweCkgb24gaXRlbVxyXG4gICAqL1xyXG4gIG1hcmdpbj86IG51bWJlcjtcclxuICAvKipcclxuICAgKiBQYWRkaW5nIGxlZnQgYW5kIHJpZ2h0IG9uIHN0YWdlIChjYW4gc2VlIG5laWdoYm91cnMpXHJcbiAgICovXHJcbiAgc3RhZ2VQYWRkaW5nPzogbnVtYmVyO1xyXG5cclxuICAvKipcclxuICAgKiBNZXJnZSBpdGVtcy4gV29ya3Mgd2l0aCBASW5wdXQgb3B0aW9uICdkYXRhTWVyZ2UnIG9mIENhcm91c2VsU2xpZGVEaXJlY3RpdmUuXHJcbiAgICovXHJcbiAgbWVyZ2U/OiBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIEZpdCBtZXJnZWQgaXRlbXMgaWYgc2NyZWVuIGlzIHNtYWxsZXIgdGhhbiBpdGVtcyB2YWx1ZS5cclxuICAgKi9cclxuICBtZXJnZUZpdD86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogU2V0IG5vbiBncmlkIGNvbnRlbnQuICBXb3JrcyB3aXRoIEBJbnB1dCBvcHRpb24gJ3dpZHRoJyBvZiBDYXJvdXNlbFNsaWRlRGlyZWN0aXZlLlxyXG4gICAqL1xyXG4gIGF1dG9XaWR0aD86IGJvb2xlYW47XHJcblxyXG4gIC8qKlxyXG4gICAqIFN0YXJ0IHBvc2l0aW9uXHJcbiAgICovXHJcbiAgc3RhcnRQb3NpdGlvbj86IG51bWJlciB8IHN0cmluZztcclxuICAvKipcclxuICAgKiBDaGFuZ2UgZGlyZWN0aW9uIGZyb20gUmlnaHQgdG8gbGVmdFxyXG4gICAqL1xyXG4gIHJ0bD86IGJvb2xlYW47XHJcblxyXG4gIC8qKlxyXG4gICAqIFNwZWVkIENhbGN1bGF0ZSB3aGlsZSBkcmFnZ2luZ1xyXG4gICAqL1xyXG4gIHNtYXJ0U3BlZWQ/OiBudW1iZXI7XHJcbiAgLyoqXHJcbiAgICogU3BlZWQgQ2FsY3VsYXRlIHdoaWxlIGRyYWdnaW5nXHJcbiAgICovXHJcbiAgZmx1aWRTcGVlZD86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogRHJhZyBlbmQgc3BlZWRcclxuICAgKi9cclxuICBkcmFnRW5kU3BlZWQ/OiBudW1iZXIgfCBib29sZWFuO1xyXG5cclxuICAvKipcclxuICAgKiBPYmplY3QgY29udGFpbmluZyByZXNwb25zaXZlIG9wdGlvbnMuIENhbiBiZSBzZXQgdG8gZmFsc2UgdG8gcmVtb3ZlIHJlc3BvbnNpdmUgY2FwYWJpbGl0aWVzLlxyXG4gICAqL1xyXG4gIHJlc3BvbnNpdmU/OiBSZXNwb25zaXZlU2V0dGluZ3M7XHJcbiAgLyoqXHJcbiAgICogUmVzcG9uc2l2ZSByZWZyZXNoIHJhdGUuIFBlcmlvZCBvZiB0aW1lIGJldHdlZW4gZmlyaW5nICdyZXNpemUnIGV2ZW50IGFuZCByZWZyZXNoaW5nIGNhcm91c2VsLlxyXG4gICAqL1xyXG4gIHJlc3BvbnNpdmVSZWZyZXNoUmF0ZT86IG51bWJlcjtcclxuXHJcbiAgLy8gTmF2aWdhdGlvblxyXG4gIC8qKlxyXG4gICAqIFNob3cgbmV4dC9wcmV2IGJ1dHRvbnMuXHJcbiAgICovXHJcbiAgbmF2PzogYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBUZXh0IGluIG5leHQvcHJldiBidXR0b25zLiBIVE1MIGFsbG93ZWQuXHJcbiAgICovXHJcbiAgbmF2VGV4dD86IHN0cmluZ1tdO1xyXG4gIC8qKlxyXG4gICAqIE5hdmlnYXRpb24gc3BlZWRcclxuICAgKi9cclxuICBuYXZTcGVlZD86IG51bWJlciB8IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogTmF2aWdhdGlvbiBzbGlkZSBieSB4LiAncGFnZScgc3RyaW5nIGNhbiBiZSBzZXQgdG8gc2xpZGUgYnkgcGFnZS5cclxuICAgKi9cclxuICBzbGlkZUJ5PzogbnVtYmVyIHwgc3RyaW5nO1xyXG4gIC8qKlxyXG4gICAqIFNob3cgZG90cyBuYXZpZ2F0aW9uXHJcbiAgICovXHJcbiAgZG90cz86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogU2hvdyBkb3RzIGVhY2ggeCBpdGVtXHJcbiAgICovXHJcbiAgZG90c0VhY2g/OiBudW1iZXIgfCBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIFVzZWQgd2l0aCBASW5wdXQgb3B0aW9uICdkb3RDb250ZW50JyBvZiBDYXJvdXNlbFNsaWRlRGlyZWN0aXZlLlxyXG4gICAqL1xyXG4gIGRvdHNEYXRhPzogYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBQYWdpbmF0aW9uIHNwZWVkLlxyXG4gICAqL1xyXG4gIGRvdHNTcGVlZD86IG51bWJlciB8IGJvb2xlYW47XHJcblxyXG5cclxuICAvLyBBVVRPUExBWSAqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqXHJcbiAgLyoqXHJcbiAgICogQXV0b3BsYXkuXHJcbiAgICovXHJcbiAgYXV0b3BsYXk/OiBib29sZWFuO1xyXG5cclxuICAvKipcclxuICAgKiBBdXRvcGxheSBpbnRlcnZhbCB0aW1lb3V0XHJcbiAgICovXHJcbiAgYXV0b3BsYXlUaW1lb3V0PzogbnVtYmVyO1xyXG5cclxuICAvKipcclxuICAgKiBQYXVzZSBvbiBtb3VzZSBob3ZlclxyXG4gICAqL1xyXG4gIGF1dG9wbGF5SG92ZXJQYXVzZT86IGJvb2xlYW47XHJcblxyXG4gIC8qKlxyXG4gICAqIEF1dG9wbGF5IHNwZWVkXHJcbiAgICovXHJcbiAgYXV0b3BsYXlTcGVlZD86IG51bWJlciB8IGJvb2xlYW47XHJcblxyXG4gIC8vIExBWlkgTE9BRElORyAqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqXHJcbiAgLyoqXHJcbiAgICogTGF6eSBsb2FkaW5nIG9mIHNsaWRlcyB3aXRoIGltYWdlc1xyXG4gICAqL1xyXG4gIGxhenlMb2FkPzogYm9vbGVhbjtcclxuXHJcbiAgLyoqXHJcbiAgICogSW5kaWNhdGVzIGhvdyBtYW55IGl0ZW1zIChzbGlkZXMgd2l0aCBpbWFnZXMpIHRvIHRoZSByaWdodCAoYW5kIGxlZnQsIHdoZW4gbG9vcCBpcyB0cnVlKSB3aWxsIGJlIHByZS1sb2FkZWRcclxuICAgKi9cclxuICBsYXp5TG9hZEVhZ2VyPzogbnVtYmVyO1xyXG5cclxuICAvLyBBTklNQVRFICoqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKipcclxuICAvKipcclxuICAgKiBDbGFzcyBmb3IgQ1NTMyBhbmltYXRpb24gb3V0LlxyXG4gICAqL1xyXG4gIGFuaW1hdGVPdXQ/OiBzdHJpbmcgfCBib29sZWFuO1xyXG5cclxuICAvKipcclxuICAgKiBDbGFzcyBmb3IgQ1NTMyBhbmltYXRpb24gaW4uXHJcbiAgICovXHJcbiAgYW5pbWF0ZUluPzogc3RyaW5nIHwgYm9vbGVhbjtcclxuXHJcbiAgLy8gQVVUT0hFSUdIVCAqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqXHJcbiAgLyoqXHJcbiAgICogRW5hYmxlcyBjaGFuZ2luZyBjYXJvdXNlbCBhY2NvcmRpbmcgdG8gdGhlIGJpZ2dlc3QgaGVpZ2h0IG9mIGFzcyBhY3RpdmUgc2xpZGVzIHByb3ZpZGluZyBzbGlkZXMgaGF2ZSBkaWZmZXJlbnQgaGVpZ2h0XHJcbiAgICovXHJcbiAgYXV0b0hlaWdodD86IGJvb2xlYW47XHJcblxyXG4gIC8vIEhBU0ggKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKioqKipcclxuICAvKipcclxuICAgKiBFbmFibGVzIGxpc3RlbmluZyB0byB1cmwgaGFzaCAoZnJhZ21lbnQpIGNoYW5nZXNcclxuICAgKi9cclxuICBVUkxoYXNoTGlzdGVuZXI/OiBib29sZWFuO1xyXG4gIGNvbnN0cnVjdG9yKCkgeyB9XHJcbn1cclxuIl19
