@@ -1,6 +1,6 @@
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 export class SlideModel {
 }
@@ -101,4 +101,4 @@ if (false) {
      */
     SlideModel.prototype.hashFragment;
 }
-//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoic2xpZGUubW9kZWwuanMiLCJzb3VyY2VSb290Ijoibmc6Ly9uZ3gtb3dsLWNhcm91c2VsLW8vIiwic291cmNlcyI6WyJsaWIvbW9kZWxzL3NsaWRlLm1vZGVsLnRzIl0sIm5hbWVzIjpbXSwibWFwcGluZ3MiOiI7Ozs7QUFFQSxNQUFNO0NBNkZMOzs7Ozs7SUF4RkMsd0JBQVc7Ozs7O0lBS1gsOEJBQW1COzs7OztJQUtuQiw0QkFBMEI7Ozs7O0lBSzFCLCtCQUFtQjs7Ozs7SUFLbkIsMkJBQXdCOzs7OztJQUt4Qiw2QkFBMEI7Ozs7O0lBSzFCLDZCQUEwQjs7Ozs7SUFLMUIsZ0NBQXFCOzs7OztJQUtyQiw4QkFBbUI7Ozs7O0lBS25CLDBCQUFlOzs7OztJQUtmLDBCQUF1Qjs7Ozs7SUFLdkIsNkJBQWtDOzs7OztJQUtsQyxnQ0FBcUI7Ozs7O0lBS3JCLHFDQUEwQjs7Ozs7SUFJMUIsc0NBQTJCOzs7OztJQUkzQix3Q0FBNkI7Ozs7O0lBSTdCLHlDQUE4Qjs7Ozs7SUFLOUIsaUNBQXFCOzs7OztJQUtyQixrQ0FBc0IiLCJzb3VyY2VzQ29udGVudCI6WyJpbXBvcnQgeyBUZW1wbGF0ZVJlZiB9IGZyb20gXCJAYW5ndWxhci9jb3JlXCI7XHJcblxyXG5leHBvcnQgY2xhc3MgU2xpZGVNb2RlbCB7XHJcblxyXG4gIC8qKlxyXG4gICAqIElkIG9mIHNsaWRlXHJcbiAgICovXHJcbiAgaWQ6IHN0cmluZztcclxuXHJcbiAgLyoqXHJcbiAgICogQWN0aXZlIHN0YXRlIG9mIHNsaWRlLiBJZiB0cnVlIHNsaWRlIGdldHMgY3NzLWNsYXNzIC5hY3RpdmVcclxuICAgKi9cclxuICBpc0FjdGl2ZT86IGJvb2xlYW47XHJcblxyXG4gIC8qKlxyXG4gICAqIFRlbXBsYXRlUmVmIG9mIHNsaWRlLiBJbiBvdGhlciB3b3JkcyBpdHMgaHRtbC1tYXJrdXBcclxuICAgKi9cclxuICB0cGxSZWY/OiBUZW1wbGF0ZVJlZjxhbnk+O1xyXG5cclxuICAvKipcclxuICAgKiBOdW1iZXIgb2YgZ3JpZCBwYXJ0cyB0byBiZSB1c2VkXHJcbiAgICovXHJcbiAgZGF0YU1lcmdlPzogbnVtYmVyO1xyXG5cclxuICAvKipcclxuICAgKiBXaWR0aCBvZiBzbGlkZVxyXG4gICAqL1xyXG4gIHdpZHRoPzogbnVtYmVyIHwgc3RyaW5nO1xyXG5cclxuICAvKipcclxuICAgKiBDc3MtcnVsZSAnbWFyZ2luLWxlZnQnXHJcbiAgICovXHJcbiAgbWFyZ2luTD86IG51bWJlciB8IHN0cmluZztcclxuXHJcbiAgLyoqXHJcbiAgICogQ3NzLXJ1bGUgJ21hcmdpbi1yaWdodCdcclxuICAgKi9cclxuICBtYXJnaW5SPzogbnVtYmVyIHwgc3RyaW5nO1xyXG5cclxuICAvKipcclxuICAgKiBNYWtlIHNsaWRlIHRvIGJlIG9uIGNlbnRlciBvZiB0aGUgY2Fyb3VzZWxcclxuICAgKi9cclxuICBpc0NlbnRlcmVkPzogYm9vbGVhbjtcclxuXHJcbiAgLyoqXHJcbiAgICogQ2xvbmVkIHNsaWRlLiBJdCdzIGJlaW5nIHVzZWQgd2hlbiAnbG9vcCc9dHJ1ZVxyXG4gICAqL1xyXG4gIGlzQ2xvbmVkPzogYm9vbGVhbjtcclxuXHJcbiAgLyoqXHJcbiAgICogSW5kaWNhdGVzIHdoZXRoZXIgc2xpZGUgc2hvdWxkIGJlIGxhenkgbG9hZGVkXHJcbiAgICovXHJcbiAgbG9hZD86IGJvb2xlYW47XHJcblxyXG4gIC8qKlxyXG4gICAqIENzcy1ydWxlICdsZWZ0J1xyXG4gICAqL1xyXG4gIGxlZnQ/OiBudW1iZXIgfCBzdHJpbmc7XHJcblxyXG4gIC8qKlxyXG4gICAqIENoYW5nZWFibGUgY2xhc3NlcyBvZiBzbGlkZVxyXG4gICAqL1xyXG4gIGNsYXNzZXM/OiB7W2tleTpzdHJpbmddOiBib29sZWFufTtcclxuXHJcbiAgLyoqXHJcbiAgICogU2hvd3Mgd2hldGhlciBzbGlkZSBjb3VsZCBiZSBhbmltYXRlZCBhbmQgY291bGQgaGF2ZSBjc3MtY2xhc3MgJy5hbmltYXRlZCdcclxuICAgKi9cclxuICBpc0FuaW1hdGVkPzogYm9vbGVhbjtcclxuXHJcbiAgLyoqXHJcbiAgICogU2hvd3Mgd2hldGhlciBzbGlkZSBjb3VsZCBiZSBhbmltYXRlZC1pbiBhbmQgY291bGQgaGF2ZSBjc3MtY2xhc3MgJy5vd2wtYW5pbWF0ZWQtaW4nXHJcbiAgICovXHJcbiAgaXNEZWZBbmltYXRlZEluPzogYm9vbGVhbjtcclxuICAvKipcclxuICAgKiBTaG93cyB3aGV0aGVyIHNsaWRlIGNvdWxkIGJlIGFuaW1hdGVkLW91dCBhbmQgY291bGQgaGF2ZSBjc3MtY2xhc3MgJy5vd2wtYW5pbWF0ZWQtb3V0J1xyXG4gICAqL1xyXG4gIGlzRGVmQW5pbWF0ZWRPdXQ/OiBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIFNob3dzIHdoZXRoZXIgc2xpZGUgY291bGQgYmUgYW5pbWF0ZWQtaW4gYW5kIGNvdWxkIGhhdmUgYW5pbWF0aW9uIGNzcy1jbGFzcyBkZWZpbmVkIGJ5IHVzZXJcclxuICAgKi9cclxuICBpc0N1c3RvbUFuaW1hdGVkSW4/OiBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIFNob3dzIHdoZXRoZXIgc2xpZGUgY291bGQgYmUgYW5pbWF0ZWQtb3V0IGFuZCBjb3VsZCBoYXZlIGFuaW1hdGlvbiBjc3MtY2xhc3MgZGVmaW5lZCBieSB1c2VyXHJcbiAgICovXHJcbiAgaXNDdXN0b21BbmltYXRlZE91dD86IGJvb2xlYW47XHJcblxyXG4gIC8qKlxyXG4gICAqIFN0YXRlIGZvciBkZWZpbmluZyB0aGUgaGVpZ2h0IG9mIHNsaWRlLkl0J3MgdmFsdWVzIGNvdWxkIGJlICdmdWxsJyBhbmQgJ251bGxlZCcuICdGdWxsJyBzZXRzIGNzcy1oZWlnaHQgdG8gJ2F1dG8nLCAnbnVsbGVkJyBzZXRzIGhlaWdodCB0byAnMCcuXHJcbiAgICovXHJcbiAgaGVpZ2h0U3RhdGU/OiBzdHJpbmc7XHJcblxyXG4gIC8qKlxyXG4gICAqIEhhc2ggKGZyYWdtZW50KSBvZiB1cmwgd2hpY2ggY29ycmVzcG9uZHMgdG8gc2xpZGVcclxuICAgKi9cclxuICBoYXNoRnJhZ21lbnQ/OiBzdHJpbmc7XHJcbn0iXX0=
+//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoic2xpZGUubW9kZWwuanMiLCJzb3VyY2VSb290Ijoibmc6Ly9uZ3gtb3dsLWNhcm91c2VsLW8vIiwic291cmNlcyI6WyJsaWIvbW9kZWxzL3NsaWRlLm1vZGVsLnRzIl0sIm5hbWVzIjpbXSwibWFwcGluZ3MiOiI7Ozs7QUFFQSxNQUFNLE9BQU8sVUFBVTtDQTZGdEI7Ozs7OztJQXhGQyx3QkFBVzs7Ozs7SUFLWCw4QkFBbUI7Ozs7O0lBS25CLDRCQUEwQjs7Ozs7SUFLMUIsK0JBQW1COzs7OztJQUtuQiwyQkFBd0I7Ozs7O0lBS3hCLDZCQUEwQjs7Ozs7SUFLMUIsNkJBQTBCOzs7OztJQUsxQixnQ0FBcUI7Ozs7O0lBS3JCLDhCQUFtQjs7Ozs7SUFLbkIsMEJBQWU7Ozs7O0lBS2YsMEJBQXVCOzs7OztJQUt2Qiw2QkFBa0M7Ozs7O0lBS2xDLGdDQUFxQjs7Ozs7SUFLckIscUNBQTBCOzs7OztJQUkxQixzQ0FBMkI7Ozs7O0lBSTNCLHdDQUE2Qjs7Ozs7SUFJN0IseUNBQThCOzs7OztJQUs5QixpQ0FBcUI7Ozs7O0lBS3JCLGtDQUFzQiIsInNvdXJjZXNDb250ZW50IjpbImltcG9ydCB7IFRlbXBsYXRlUmVmIH0gZnJvbSBcIkBhbmd1bGFyL2NvcmVcIjtcclxuXHJcbmV4cG9ydCBjbGFzcyBTbGlkZU1vZGVsIHtcclxuXHJcbiAgLyoqXHJcbiAgICogSWQgb2Ygc2xpZGVcclxuICAgKi9cclxuICBpZDogc3RyaW5nO1xyXG5cclxuICAvKipcclxuICAgKiBBY3RpdmUgc3RhdGUgb2Ygc2xpZGUuIElmIHRydWUgc2xpZGUgZ2V0cyBjc3MtY2xhc3MgLmFjdGl2ZVxyXG4gICAqL1xyXG4gIGlzQWN0aXZlPzogYm9vbGVhbjtcclxuXHJcbiAgLyoqXHJcbiAgICogVGVtcGxhdGVSZWYgb2Ygc2xpZGUuIEluIG90aGVyIHdvcmRzIGl0cyBodG1sLW1hcmt1cFxyXG4gICAqL1xyXG4gIHRwbFJlZj86IFRlbXBsYXRlUmVmPGFueT47XHJcblxyXG4gIC8qKlxyXG4gICAqIE51bWJlciBvZiBncmlkIHBhcnRzIHRvIGJlIHVzZWRcclxuICAgKi9cclxuICBkYXRhTWVyZ2U/OiBudW1iZXI7XHJcblxyXG4gIC8qKlxyXG4gICAqIFdpZHRoIG9mIHNsaWRlXHJcbiAgICovXHJcbiAgd2lkdGg/OiBudW1iZXIgfCBzdHJpbmc7XHJcblxyXG4gIC8qKlxyXG4gICAqIENzcy1ydWxlICdtYXJnaW4tbGVmdCdcclxuICAgKi9cclxuICBtYXJnaW5MPzogbnVtYmVyIHwgc3RyaW5nO1xyXG5cclxuICAvKipcclxuICAgKiBDc3MtcnVsZSAnbWFyZ2luLXJpZ2h0J1xyXG4gICAqL1xyXG4gIG1hcmdpblI/OiBudW1iZXIgfCBzdHJpbmc7XHJcblxyXG4gIC8qKlxyXG4gICAqIE1ha2Ugc2xpZGUgdG8gYmUgb24gY2VudGVyIG9mIHRoZSBjYXJvdXNlbFxyXG4gICAqL1xyXG4gIGlzQ2VudGVyZWQ/OiBib29sZWFuO1xyXG5cclxuICAvKipcclxuICAgKiBDbG9uZWQgc2xpZGUuIEl0J3MgYmVpbmcgdXNlZCB3aGVuICdsb29wJz10cnVlXHJcbiAgICovXHJcbiAgaXNDbG9uZWQ/OiBib29sZWFuO1xyXG5cclxuICAvKipcclxuICAgKiBJbmRpY2F0ZXMgd2hldGhlciBzbGlkZSBzaG91bGQgYmUgbGF6eSBsb2FkZWRcclxuICAgKi9cclxuICBsb2FkPzogYm9vbGVhbjtcclxuXHJcbiAgLyoqXHJcbiAgICogQ3NzLXJ1bGUgJ2xlZnQnXHJcbiAgICovXHJcbiAgbGVmdD86IG51bWJlciB8IHN0cmluZztcclxuXHJcbiAgLyoqXHJcbiAgICogQ2hhbmdlYWJsZSBjbGFzc2VzIG9mIHNsaWRlXHJcbiAgICovXHJcbiAgY2xhc3Nlcz86IHtba2V5OnN0cmluZ106IGJvb2xlYW59O1xyXG5cclxuICAvKipcclxuICAgKiBTaG93cyB3aGV0aGVyIHNsaWRlIGNvdWxkIGJlIGFuaW1hdGVkIGFuZCBjb3VsZCBoYXZlIGNzcy1jbGFzcyAnLmFuaW1hdGVkJ1xyXG4gICAqL1xyXG4gIGlzQW5pbWF0ZWQ/OiBib29sZWFuO1xyXG5cclxuICAvKipcclxuICAgKiBTaG93cyB3aGV0aGVyIHNsaWRlIGNvdWxkIGJlIGFuaW1hdGVkLWluIGFuZCBjb3VsZCBoYXZlIGNzcy1jbGFzcyAnLm93bC1hbmltYXRlZC1pbidcclxuICAgKi9cclxuICBpc0RlZkFuaW1hdGVkSW4/OiBib29sZWFuO1xyXG4gIC8qKlxyXG4gICAqIFNob3dzIHdoZXRoZXIgc2xpZGUgY291bGQgYmUgYW5pbWF0ZWQtb3V0IGFuZCBjb3VsZCBoYXZlIGNzcy1jbGFzcyAnLm93bC1hbmltYXRlZC1vdXQnXHJcbiAgICovXHJcbiAgaXNEZWZBbmltYXRlZE91dD86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogU2hvd3Mgd2hldGhlciBzbGlkZSBjb3VsZCBiZSBhbmltYXRlZC1pbiBhbmQgY291bGQgaGF2ZSBhbmltYXRpb24gY3NzLWNsYXNzIGRlZmluZWQgYnkgdXNlclxyXG4gICAqL1xyXG4gIGlzQ3VzdG9tQW5pbWF0ZWRJbj86IGJvb2xlYW47XHJcbiAgLyoqXHJcbiAgICogU2hvd3Mgd2hldGhlciBzbGlkZSBjb3VsZCBiZSBhbmltYXRlZC1vdXQgYW5kIGNvdWxkIGhhdmUgYW5pbWF0aW9uIGNzcy1jbGFzcyBkZWZpbmVkIGJ5IHVzZXJcclxuICAgKi9cclxuICBpc0N1c3RvbUFuaW1hdGVkT3V0PzogYm9vbGVhbjtcclxuXHJcbiAgLyoqXHJcbiAgICogU3RhdGUgZm9yIGRlZmluaW5nIHRoZSBoZWlnaHQgb2Ygc2xpZGUuSXQncyB2YWx1ZXMgY291bGQgYmUgJ2Z1bGwnIGFuZCAnbnVsbGVkJy4gJ0Z1bGwnIHNldHMgY3NzLWhlaWdodCB0byAnYXV0bycsICdudWxsZWQnIHNldHMgaGVpZ2h0IHRvICcwJy5cclxuICAgKi9cclxuICBoZWlnaHRTdGF0ZT86IHN0cmluZztcclxuXHJcbiAgLyoqXHJcbiAgICogSGFzaCAoZnJhZ21lbnQpIG9mIHVybCB3aGljaCBjb3JyZXNwb25kcyB0byBzbGlkZVxyXG4gICAqL1xyXG4gIGhhc2hGcmFnbWVudD86IHN0cmluZztcclxufSJdfQ==
