@@ -1,6 +1,6 @@
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 /**
  * Data model for single navigation button
@@ -9,9 +9,6 @@ var /**
  * Data model for single navigation button
  */
 NavButton = /** @class */ (function () {
-    /**
-     * Data model for single navigation button
-     */
     function NavButton() {
     }
     return NavButton;
@@ -39,9 +36,6 @@ var /**
  * Data model for  navigation block
  */
 NavData = /** @class */ (function () {
-    /**
-     * Data model for  navigation block
-     */
     function NavData() {
     }
     return NavData;
@@ -74,9 +68,6 @@ var /**
  * data model for single owl dot-button
  */
 OwlSingeDot = /** @class */ (function () {
-    /**
-     * data model for single owl dot-button
-     */
     function OwlSingeDot() {
     }
     return OwlSingeDot;
@@ -114,9 +105,6 @@ var /**
  * Data model for owl dot-block
  */
 DotsData = /** @class */ (function () {
-    /**
-     * Data model for owl dot-block
-     */
     function DotsData() {
     }
     return DotsData;
@@ -137,4 +125,4 @@ if (false) {
      */
     DotsData.prototype.dots;
 }
-//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoibmF2aWdhdGlvbi1kYXRhLm1vZGVscy5qcyIsInNvdXJjZVJvb3QiOiJuZzovL25neC1vd2wtY2Fyb3VzZWwtby8iLCJzb3VyY2VzIjpbImxpYi9tb2RlbHMvbmF2aWdhdGlvbi1kYXRhLm1vZGVscy50cyJdLCJuYW1lcyI6W10sIm1hcHBpbmdzIjoiOzs7Ozs7O0FBR0E7Ozs7SUFIQTs7T0FFRztJQUNIO0lBU0EsQ0FBQztJQUFELGdCQUFDO0FBQUQsQ0FBQyxBQVRELElBU0M7Ozs7Ozs7Ozs7SUFMQSw2QkFBa0I7Ozs7O0lBSWxCLDZCQUFpQjs7Ozs7QUFLbEI7Ozs7SUFIQTs7T0FFRztJQUNIO0lBYUEsQ0FBQztJQUFELGNBQUM7QUFBRCxDQUFDLEFBYkQsSUFhQzs7Ozs7Ozs7OztJQVRBLDJCQUFrQjs7Ozs7SUFJbEIsdUJBQWdCOzs7OztJQUloQix1QkFBZ0I7Ozs7O0FBS2pCOzs7O0lBSEE7O09BRUc7SUFDSDtJQWlCQSxDQUFDO0lBQUQsa0JBQUM7QUFBRCxDQUFDLEFBakJELElBaUJDOzs7Ozs7Ozs7O0lBYkEseUJBQVc7Ozs7O0lBSVgsNkJBQWdCOzs7OztJQUloQixtQ0FBc0I7Ozs7O0lBSXRCLHVDQUEyQjs7Ozs7QUFLNUI7Ozs7SUFIQTs7T0FFRztJQUNIO0lBU0EsQ0FBQztJQUFELGVBQUM7QUFBRCxDQUFDLEFBVEQsSUFTQzs7Ozs7Ozs7OztJQUxBLDRCQUFrQjs7Ozs7SUFJbEIsd0JBQW9CIiwic291cmNlc0NvbnRlbnQiOlsiLyoqXHJcbiAqIERhdGEgbW9kZWwgZm9yIHNpbmdsZSBuYXZpZ2F0aW9uIGJ1dHRvblxyXG4gKi9cclxuZXhwb3J0IGNsYXNzIE5hdkJ1dHRvbiB7XHJcblx0LyoqXHJcblx0ICogTWFrZSBidXR0b24gZGlzYWJsZWQgYnkgYXR0YWNoaW5nIGNsYXNzIC5kaXNhYmxlZDsgaXQncyBiZWluZyB1c2VkIHdoZW4gb3RwaW9uIGxvb3A9ZmFsc2VcclxuXHQgKi9cclxuXHRkaXNhYmxlZDogYm9vbGVhbjtcclxuXHQvKipcclxuXHQgKiBUZXh0IGZvciBidXR0b247IGNhbiBiZSBzaW1wbGUgdGV4dCBvciBodG1sLW1hcmt1cCAoZS5nLiA8aSBjbGFzcz1cImZhIGZhLW5leHRcIj4pXHJcblx0ICovXHJcblx0aHRtbFRleHQ6IHN0cmluZztcclxufVxyXG4vKipcclxuICogRGF0YSBtb2RlbCBmb3IgIG5hdmlnYXRpb24gYmxvY2tcclxuICovXHJcbmV4cG9ydCBjbGFzcyBOYXZEYXRhIHtcclxuXHQvKipcclxuXHQgKiBSdXJucyBvbiBvciB0dXJucyBvZmYgbmF2aWdhdGlvbiBibG9ja1xyXG5cdCAqL1xyXG5cdGRpc2FibGVkOiBib29sZWFuO1xyXG5cdC8qKlxyXG5cdCAqIE5hdmlnYXRpb24gYnV0dG9uICdwcmV2J1xyXG5cdCAqL1xyXG5cdHByZXY6IE5hdkJ1dHRvbjtcclxuXHQvKipcclxuXHQgKiBOYXZpZ2F0aW9uIGJ1dHRvbiAnbmV4dCdcclxuXHQgKi9cclxuXHRuZXh0OiBOYXZCdXR0b247XHJcbn1cclxuLyoqXHJcbiAqIGRhdGEgbW9kZWwgZm9yIHNpbmdsZSBvd2wgZG90LWJ1dHRvblxyXG4gKi9cclxuZXhwb3J0IGNsYXNzIE93bFNpbmdlRG90IHtcclxuXHQvKipcclxuXHQgKiBJZCBmb3IgZG90IGJ1dHRvblxyXG5cdCAqL1xyXG5cdGlkOiBzdHJpbmc7XHJcblx0LyoqXHJcblx0ICogTWFrZXMgZG90IGFjdGl2ZSBieSBhdHRhY2hpbmcgLmFjdGl2ZSBjbGFzcyB0byBpdFxyXG5cdCAqL1xyXG5cdGFjdGl2ZTogYm9vbGVhbjtcclxuXHQvKipcclxuXHQgKiBJbm5lciBjb250ZW50IG9mIGRvdDsgY2FuIGJlIGh0bWwtbWFya3VwXHJcblx0ICovXHJcblx0aW5uZXJDb250ZW50Pzogc3RyaW5nO1xyXG5cdC8qKlxyXG5cdCAqIEVuYWJsZWQgY3NzLWNsYXNzIHdoaWNoIGdpdmVzIHJpZ2h0IHByZXNlbnRhaW9uIG9mIGlubmVyQ29udGVudCBvZiBkb3QuXHJcblx0ICovXHJcblx0c2hvd0lubmVyQ29udGVudD86IGJvb2xlYW47XHJcbn1cclxuLyoqXHJcbiAqIERhdGEgbW9kZWwgZm9yIG93bCBkb3QtYmxvY2tcclxuICovXHJcbmV4cG9ydCBjbGFzcyBEb3RzRGF0YSB7XHJcblx0LyoqXHJcblx0ICogVHVybnMgb24gb3IgdHVybnMgb2ZmIG5hdmlnYXRpb24gYmxvY2tcclxuXHQgKi9cclxuXHRkaXNhYmxlZDogYm9vbGVhbjtcclxuXHQvKipcclxuXHQgKiBBcnJheSBvZiBkb3RzXHJcblx0ICovXHJcblx0ZG90czogT3dsU2luZ2VEb3RbXTtcclxufSJdfQ==
+//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoibmF2aWdhdGlvbi1kYXRhLm1vZGVscy5qcyIsInNvdXJjZVJvb3QiOiJuZzovL25neC1vd2wtY2Fyb3VzZWwtby8iLCJzb3VyY2VzIjpbImxpYi9tb2RlbHMvbmF2aWdhdGlvbi1kYXRhLm1vZGVscy50cyJdLCJuYW1lcyI6W10sIm1hcHBpbmdzIjoiOzs7Ozs7O0FBR0E7Ozs7SUFBQTtJQVNBLENBQUM7SUFBRCxnQkFBQztBQUFELENBQUMsQUFURCxJQVNDOzs7Ozs7Ozs7O0lBTEEsNkJBQWtCOzs7OztJQUlsQiw2QkFBaUI7Ozs7O0FBS2xCOzs7O0lBQUE7SUFhQSxDQUFDO0lBQUQsY0FBQztBQUFELENBQUMsQUFiRCxJQWFDOzs7Ozs7Ozs7O0lBVEEsMkJBQWtCOzs7OztJQUlsQix1QkFBZ0I7Ozs7O0lBSWhCLHVCQUFnQjs7Ozs7QUFLakI7Ozs7SUFBQTtJQWlCQSxDQUFDO0lBQUQsa0JBQUM7QUFBRCxDQUFDLEFBakJELElBaUJDOzs7Ozs7Ozs7O0lBYkEseUJBQVc7Ozs7O0lBSVgsNkJBQWdCOzs7OztJQUloQixtQ0FBc0I7Ozs7O0lBSXRCLHVDQUEyQjs7Ozs7QUFLNUI7Ozs7SUFBQTtJQVNBLENBQUM7SUFBRCxlQUFDO0FBQUQsQ0FBQyxBQVRELElBU0M7Ozs7Ozs7Ozs7SUFMQSw0QkFBa0I7Ozs7O0lBSWxCLHdCQUFvQiIsInNvdXJjZXNDb250ZW50IjpbIi8qKlxyXG4gKiBEYXRhIG1vZGVsIGZvciBzaW5nbGUgbmF2aWdhdGlvbiBidXR0b25cclxuICovXHJcbmV4cG9ydCBjbGFzcyBOYXZCdXR0b24ge1xyXG5cdC8qKlxyXG5cdCAqIE1ha2UgYnV0dG9uIGRpc2FibGVkIGJ5IGF0dGFjaGluZyBjbGFzcyAuZGlzYWJsZWQ7IGl0J3MgYmVpbmcgdXNlZCB3aGVuIG90cGlvbiBsb29wPWZhbHNlXHJcblx0ICovXHJcblx0ZGlzYWJsZWQ6IGJvb2xlYW47XHJcblx0LyoqXHJcblx0ICogVGV4dCBmb3IgYnV0dG9uOyBjYW4gYmUgc2ltcGxlIHRleHQgb3IgaHRtbC1tYXJrdXAgKGUuZy4gPGkgY2xhc3M9XCJmYSBmYS1uZXh0XCI+KVxyXG5cdCAqL1xyXG5cdGh0bWxUZXh0OiBzdHJpbmc7XHJcbn1cclxuLyoqXHJcbiAqIERhdGEgbW9kZWwgZm9yICBuYXZpZ2F0aW9uIGJsb2NrXHJcbiAqL1xyXG5leHBvcnQgY2xhc3MgTmF2RGF0YSB7XHJcblx0LyoqXHJcblx0ICogUnVybnMgb24gb3IgdHVybnMgb2ZmIG5hdmlnYXRpb24gYmxvY2tcclxuXHQgKi9cclxuXHRkaXNhYmxlZDogYm9vbGVhbjtcclxuXHQvKipcclxuXHQgKiBOYXZpZ2F0aW9uIGJ1dHRvbiAncHJldidcclxuXHQgKi9cclxuXHRwcmV2OiBOYXZCdXR0b247XHJcblx0LyoqXHJcblx0ICogTmF2aWdhdGlvbiBidXR0b24gJ25leHQnXHJcblx0ICovXHJcblx0bmV4dDogTmF2QnV0dG9uO1xyXG59XHJcbi8qKlxyXG4gKiBkYXRhIG1vZGVsIGZvciBzaW5nbGUgb3dsIGRvdC1idXR0b25cclxuICovXHJcbmV4cG9ydCBjbGFzcyBPd2xTaW5nZURvdCB7XHJcblx0LyoqXHJcblx0ICogSWQgZm9yIGRvdCBidXR0b25cclxuXHQgKi9cclxuXHRpZDogc3RyaW5nO1xyXG5cdC8qKlxyXG5cdCAqIE1ha2VzIGRvdCBhY3RpdmUgYnkgYXR0YWNoaW5nIC5hY3RpdmUgY2xhc3MgdG8gaXRcclxuXHQgKi9cclxuXHRhY3RpdmU6IGJvb2xlYW47XHJcblx0LyoqXHJcblx0ICogSW5uZXIgY29udGVudCBvZiBkb3Q7IGNhbiBiZSBodG1sLW1hcmt1cFxyXG5cdCAqL1xyXG5cdGlubmVyQ29udGVudD86IHN0cmluZztcclxuXHQvKipcclxuXHQgKiBFbmFibGVkIGNzcy1jbGFzcyB3aGljaCBnaXZlcyByaWdodCBwcmVzZW50YWlvbiBvZiBpbm5lckNvbnRlbnQgb2YgZG90LlxyXG5cdCAqL1xyXG5cdHNob3dJbm5lckNvbnRlbnQ/OiBib29sZWFuO1xyXG59XHJcbi8qKlxyXG4gKiBEYXRhIG1vZGVsIGZvciBvd2wgZG90LWJsb2NrXHJcbiAqL1xyXG5leHBvcnQgY2xhc3MgRG90c0RhdGEge1xyXG5cdC8qKlxyXG5cdCAqIFR1cm5zIG9uIG9yIHR1cm5zIG9mZiBuYXZpZ2F0aW9uIGJsb2NrXHJcblx0ICovXHJcblx0ZGlzYWJsZWQ6IGJvb2xlYW47XHJcblx0LyoqXHJcblx0ICogQXJyYXkgb2YgZG90c1xyXG5cdCAqL1xyXG5cdGRvdHM6IE93bFNpbmdlRG90W107XHJcbn0iXX0=
