@@ -1,6 +1,6 @@
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 /**
  * Data model for managing classes of .owl-carousel DOM element
@@ -9,9 +9,6 @@ var /**
  * Data model for managing classes of .owl-carousel DOM element
  */
 OwlDOMData = /** @class */ (function () {
-    /**
-     * Data model for managing classes of .owl-carousel DOM element
-     */
     function OwlDOMData() {
     }
     return OwlDOMData;
@@ -62,4 +59,4 @@ if (false) {
      */
     OwlDOMData.prototype.isGrab;
 }
-//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoib3dsRE9NLWRhdGEubW9kZWwuanMiLCJzb3VyY2VSb290Ijoibmc6Ly9uZ3gtb3dsLWNhcm91c2VsLW8vIiwic291cmNlcyI6WyJsaWIvbW9kZWxzL293bERPTS1kYXRhLm1vZGVsLnRzIl0sIm5hbWVzIjpbXSwibWFwcGluZ3MiOiI7Ozs7Ozs7QUFHQTs7OztJQUhBOztPQUVHO0lBQ0g7SUFpQ0EsQ0FBQztJQUFELGlCQUFDO0FBQUQsQ0FBQyxBQWpDRCxJQWlDQzs7Ozs7Ozs7OztJQTdCQSx5QkFBYTs7Ozs7SUFJYixrQ0FBc0I7Ozs7O0lBSXRCLGlDQUFxQjs7Ozs7SUFJckIsOEJBQWtCOzs7OztJQUlsQiwrQkFBbUI7Ozs7O0lBSW5CLHFDQUF5Qjs7Ozs7SUFJekIscUNBQXlCOzs7OztJQUl6Qiw0QkFBZ0IiLCJzb3VyY2VzQ29udGVudCI6WyIvKipcclxuICogRGF0YSBtb2RlbCBmb3IgbWFuYWdpbmcgY2xhc3NlcyBvZiAub3dsLWNhcm91c2VsIERPTSBlbGVtZW50XHJcbiAqL1xyXG5leHBvcnQgY2xhc3MgT3dsRE9NRGF0YSB7XHJcblx0LyoqXHJcblx0ICogRGVmaW5lcyB3aGV0aGVyIHRvIHNldCBjbGFzcyAub3dsLXJ0bCBvciBub3RcclxuXHQgKi9cclxuXHRydGw6IGJvb2xlYW47XHJcblx0LyoqXHJcblx0ICogRGVmaW5lcyB3aGV0aGVyIHRvIHNldCBjbGFzcyAub3dsLXJlc3BvbnNpdmUgb3Igbm90XHJcblx0ICovXHJcblx0aXNSZXNwb25zaXZlOiBib29sZWFuO1xyXG5cdC8qKlxyXG5cdCAqIERlZmluZXMgd2hldGhlciB0byBzZXQgY2xhc3MgLm93bC1yZWZyZXNoZWQgb3Igbm90XHJcblx0ICovXHJcblx0aXNSZWZyZXNoZWQ6IGJvb2xlYW47XHJcblx0LyoqXHJcblx0ICogRGVmaW5lcyB3aGV0aGVyIHRvIHNldCBjbGFzcyAub3dsLWxvYWRlZCBvciBub3RcclxuXHQqL1xyXG5cdGlzTG9hZGVkOiBib29sZWFuO1xyXG5cdC8qKlxyXG5cdCAqIERlZmluZXMgd2hldGhlciB0byBzZXQgY2xhc3MgLm93bC1sb2FkaW5nIG9yIG5vdFxyXG5cdCAqL1xyXG5cdGlzTG9hZGluZzogYm9vbGVhbjtcclxuXHQvKipcclxuXHQgKiBEZWZpbmVzIHdoZXRoZXIgdG8gc2V0IGNsYXNzIC5vd2wtZHJhZyBvciBub3QgYW5kIG1ha2VzIGNhcm91c2VsIGRyYWdnYWJsZSBieSBtb3VzZSBtb3ZpbmdcclxuXHQgKi9cclxuXHRpc01vdXNlRHJhZ2FibGU6IGJvb2xlYW47XHJcblx0LyoqXHJcblx0ICogTWFrZXMgY2Fyb3VzZWwgZHJhZ2dhYmxlIGJ5IHRvdWNoIG1vdmluZ1xyXG5cdCAqL1xyXG5cdGlzVG91Y2hEcmFnYWJsZTogYm9vbGVhbjtcclxuXHQvKipcclxuXHQgKiBEZWZpbmVzIHdoZXRoZXIgdG8gc2V0IGNsYXNzIC5vd2wtZ3JhYiBvciBub3RcclxuXHQgKi9cclxuXHRpc0dyYWI6IGJvb2xlYW47XHJcbn0iXX0=
+//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoib3dsRE9NLWRhdGEubW9kZWwuanMiLCJzb3VyY2VSb290Ijoibmc6Ly9uZ3gtb3dsLWNhcm91c2VsLW8vIiwic291cmNlcyI6WyJsaWIvbW9kZWxzL293bERPTS1kYXRhLm1vZGVsLnRzIl0sIm5hbWVzIjpbXSwibWFwcGluZ3MiOiI7Ozs7Ozs7QUFHQTs7OztJQUFBO0lBaUNBLENBQUM7SUFBRCxpQkFBQztBQUFELENBQUMsQUFqQ0QsSUFpQ0M7Ozs7Ozs7Ozs7SUE3QkEseUJBQWE7Ozs7O0lBSWIsa0NBQXNCOzs7OztJQUl0QixpQ0FBcUI7Ozs7O0lBSXJCLDhCQUFrQjs7Ozs7SUFJbEIsK0JBQW1COzs7OztJQUluQixxQ0FBeUI7Ozs7O0lBSXpCLHFDQUF5Qjs7Ozs7SUFJekIsNEJBQWdCIiwic291cmNlc0NvbnRlbnQiOlsiLyoqXHJcbiAqIERhdGEgbW9kZWwgZm9yIG1hbmFnaW5nIGNsYXNzZXMgb2YgLm93bC1jYXJvdXNlbCBET00gZWxlbWVudFxyXG4gKi9cclxuZXhwb3J0IGNsYXNzIE93bERPTURhdGEge1xyXG5cdC8qKlxyXG5cdCAqIERlZmluZXMgd2hldGhlciB0byBzZXQgY2xhc3MgLm93bC1ydGwgb3Igbm90XHJcblx0ICovXHJcblx0cnRsOiBib29sZWFuO1xyXG5cdC8qKlxyXG5cdCAqIERlZmluZXMgd2hldGhlciB0byBzZXQgY2xhc3MgLm93bC1yZXNwb25zaXZlIG9yIG5vdFxyXG5cdCAqL1xyXG5cdGlzUmVzcG9uc2l2ZTogYm9vbGVhbjtcclxuXHQvKipcclxuXHQgKiBEZWZpbmVzIHdoZXRoZXIgdG8gc2V0IGNsYXNzIC5vd2wtcmVmcmVzaGVkIG9yIG5vdFxyXG5cdCAqL1xyXG5cdGlzUmVmcmVzaGVkOiBib29sZWFuO1xyXG5cdC8qKlxyXG5cdCAqIERlZmluZXMgd2hldGhlciB0byBzZXQgY2xhc3MgLm93bC1sb2FkZWQgb3Igbm90XHJcblx0Ki9cclxuXHRpc0xvYWRlZDogYm9vbGVhbjtcclxuXHQvKipcclxuXHQgKiBEZWZpbmVzIHdoZXRoZXIgdG8gc2V0IGNsYXNzIC5vd2wtbG9hZGluZyBvciBub3RcclxuXHQgKi9cclxuXHRpc0xvYWRpbmc6IGJvb2xlYW47XHJcblx0LyoqXHJcblx0ICogRGVmaW5lcyB3aGV0aGVyIHRvIHNldCBjbGFzcyAub3dsLWRyYWcgb3Igbm90IGFuZCBtYWtlcyBjYXJvdXNlbCBkcmFnZ2FibGUgYnkgbW91c2UgbW92aW5nXHJcblx0ICovXHJcblx0aXNNb3VzZURyYWdhYmxlOiBib29sZWFuO1xyXG5cdC8qKlxyXG5cdCAqIE1ha2VzIGNhcm91c2VsIGRyYWdnYWJsZSBieSB0b3VjaCBtb3ZpbmdcclxuXHQgKi9cclxuXHRpc1RvdWNoRHJhZ2FibGU6IGJvb2xlYW47XHJcblx0LyoqXHJcblx0ICogRGVmaW5lcyB3aGV0aGVyIHRvIHNldCBjbGFzcyAub3dsLWdyYWIgb3Igbm90XHJcblx0ICovXHJcblx0aXNHcmFiOiBib29sZWFuO1xyXG59Il19
