@@ -1,6 +1,6 @@
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes,extraRequire,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ * @suppress {checkTypes,extraRequire,missingReturn,uselessCode} checked by tsc
  */
 /**
  * Data model for managing classes of .owl-stage DOM element
@@ -9,9 +9,6 @@ var /**
  * Data model for managing classes of .owl-stage DOM element
  */
 StageData = /** @class */ (function () {
-    /**
-     * Data model for managing classes of .owl-stage DOM element
-     */
     function StageData() {
     }
     return StageData;
@@ -47,4 +44,4 @@ if (false) {
      */
     StageData.prototype.paddingR;
 }
-//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoic3RhZ2UtZGF0YS5tb2RlbC5qcyIsInNvdXJjZVJvb3QiOiJuZzovL25neC1vd2wtY2Fyb3VzZWwtby8iLCJzb3VyY2VzIjpbImxpYi9tb2RlbHMvc3RhZ2UtZGF0YS5tb2RlbC50cyJdLCJuYW1lcyI6W10sIm1hcHBpbmdzIjoiOzs7Ozs7O0FBR0E7Ozs7SUFIQTs7T0FFRztJQUNIO0lBcUJBLENBQUM7SUFBRCxnQkFBQztBQUFELENBQUMsQUFyQkQsSUFxQkM7Ozs7Ozs7Ozs7SUFqQkEsOEJBQWtCOzs7OztJQUlsQiwrQkFBbUI7Ozs7O0lBSW5CLDBCQUF1Qjs7Ozs7SUFJdkIsNkJBQTBCOzs7OztJQUkxQiw2QkFBMEIiLCJzb3VyY2VzQ29udGVudCI6WyIvKipcclxuICogRGF0YSBtb2RlbCBmb3IgbWFuYWdpbmcgY2xhc3NlcyBvZiAub3dsLXN0YWdlIERPTSBlbGVtZW50XHJcbiAqL1xyXG5leHBvcnQgY2xhc3MgU3RhZ2VEYXRhIHtcclxuXHQvKipcclxuXHQgKiBEZXRlcm1pbmVzIGNzcy1ydWxlICd0cmFuc2Zvcm0nXHJcblx0ICovXHJcblx0dHJhbnNmb3JtOiBzdHJpbmc7XHJcblx0LyoqXHJcblx0ICogIERldGVybWluZXMgY3NzLXJ1bGUgJ3RyYW5zaXRpb24nXHJcbiAgICovXHJcblx0dHJhbnNpdGlvbjogc3RyaW5nO1xyXG5cdC8qKlxyXG5cdCAqICBEZXRlcm1pbmVzIGNzcy1ydWxlICd3aWR0aCdcclxuICAgKi9cclxuXHR3aWR0aDogbnVtYmVyIHwgc3RyaW5nO1xyXG5cdC8qKlxyXG5cdCAqICBEZXRlcm1pbmVzIGNzcy1ydWxlICdwYWRkaW5nLWxlZnQnXHJcbiAgICovXHJcblx0cGFkZGluZ0w6IG51bWJlciB8IHN0cmluZztcclxuXHQvKipcclxuXHQgKiAgRGV0ZXJtaW5lcyBjc3MtcnVsZSAncGFkZGluZy1yaWdodCdcclxuICAgKi9cclxuXHRwYWRkaW5nUjogbnVtYmVyIHwgc3RyaW5nO1xyXG59Il19
+//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoic3RhZ2UtZGF0YS5tb2RlbC5qcyIsInNvdXJjZVJvb3QiOiJuZzovL25neC1vd2wtY2Fyb3VzZWwtby8iLCJzb3VyY2VzIjpbImxpYi9tb2RlbHMvc3RhZ2UtZGF0YS5tb2RlbC50cyJdLCJuYW1lcyI6W10sIm1hcHBpbmdzIjoiOzs7Ozs7O0FBR0E7Ozs7SUFBQTtJQXFCQSxDQUFDO0lBQUQsZ0JBQUM7QUFBRCxDQUFDLEFBckJELElBcUJDOzs7Ozs7Ozs7O0lBakJBLDhCQUFrQjs7Ozs7SUFJbEIsK0JBQW1COzs7OztJQUluQiwwQkFBdUI7Ozs7O0lBSXZCLDZCQUEwQjs7Ozs7SUFJMUIsNkJBQTBCIiwic291cmNlc0NvbnRlbnQiOlsiLyoqXHJcbiAqIERhdGEgbW9kZWwgZm9yIG1hbmFnaW5nIGNsYXNzZXMgb2YgLm93bC1zdGFnZSBET00gZWxlbWVudFxyXG4gKi9cclxuZXhwb3J0IGNsYXNzIFN0YWdlRGF0YSB7XHJcblx0LyoqXHJcblx0ICogRGV0ZXJtaW5lcyBjc3MtcnVsZSAndHJhbnNmb3JtJ1xyXG5cdCAqL1xyXG5cdHRyYW5zZm9ybTogc3RyaW5nO1xyXG5cdC8qKlxyXG5cdCAqICBEZXRlcm1pbmVzIGNzcy1ydWxlICd0cmFuc2l0aW9uJ1xyXG4gICAqL1xyXG5cdHRyYW5zaXRpb246IHN0cmluZztcclxuXHQvKipcclxuXHQgKiAgRGV0ZXJtaW5lcyBjc3MtcnVsZSAnd2lkdGgnXHJcbiAgICovXHJcblx0d2lkdGg6IG51bWJlciB8IHN0cmluZztcclxuXHQvKipcclxuXHQgKiAgRGV0ZXJtaW5lcyBjc3MtcnVsZSAncGFkZGluZy1sZWZ0J1xyXG4gICAqL1xyXG5cdHBhZGRpbmdMOiBudW1iZXIgfCBzdHJpbmc7XHJcblx0LyoqXHJcblx0ICogIERldGVybWluZXMgY3NzLXJ1bGUgJ3BhZGRpbmctcmlnaHQnXHJcbiAgICovXHJcblx0cGFkZGluZ1I6IG51bWJlciB8IHN0cmluZztcclxufSJdfQ==
