@@ -91,41 +91,41 @@ export declare class StageComponent implements OnInit, OnDestroy {
        * @todo #261
        * @param event - The event arguments.
        */
-    private _onDragStart(event);
+    private _onDragStart;
     /**
      * Attaches listeners to `touchmove` and `mousemove` events; initiates updating carousel after starting dragging
      * @param event event objech of mouse or touch event
      */
-    private _oneMouseTouchMove(event);
+    private _oneMouseTouchMove;
     /**
      * Attaches handler to HTMLAnchorElement for preventing click while carousel is being dragged
      * @param event event object
      */
-    private blockClickAnchorInDragging(event);
+    private blockClickAnchorInDragging;
     /**
      * Handles the `touchmove` and `mousemove` events.
      * @todo #261
      * @param event - The event arguments.
      */
-    private _onDragMove(event);
+    private _onDragMove;
     /**
      * Moves .owl-stage left-right
      * @param coordinate coordinate to be set to .owl-stage
      */
-    private _animate(coordinate);
+    private _animate;
     /**
        * Handles the `touchend` and `mouseup` events.
        * @todo #261
        * @todo Threshold for click event
        * @param event - The event arguments.
        */
-    private _onDragEnd(event);
+    private _onDragEnd;
     /**
        * Prepares data for dragging carousel. It starts after firing `touchstart` and `mousedown` events.
        * @param event - The event arguments.
        * @returns stage - object with 'x' and 'y' coordinates of .owl-stage
        */
-    private _prepareDragging(event);
+    private _prepareDragging;
     /**
      * Attaches handler for 'click' event on any element in .owl-stage in order to prevent dragging when moving of cursor is less than 3px
      */
@@ -134,35 +134,35 @@ export declare class StageComponent implements OnInit, OnDestroy {
      * Finishes dragging
      * @param event object event of 'mouseUp' of 'touchend' events
      */
-    private _finishDragging(event);
+    private _finishDragging;
     /**
        * Gets unified pointer coordinates from event.
        * @param event The `mousedown` or `touchstart` event.
        * @returns Contains `x` and `y` coordinates of current pointer position.
        */
-    private _pointer(event);
+    private _pointer;
     /**
        * Gets the difference of two vectors.
        * @param first The first vector.
        * @param second- The second vector.
        * @returns The difference.
        */
-    private _difference(firstC, second);
+    private _difference;
     /**
        * Checks whether the carousel is in a specific state or not.
        * @param specificState The state to check.
        * @returns The flag which indicates if the carousel is busy.
        */
-    private _is(specificState);
+    private _is;
     /**
     * Enters a state.
     * @param name The state name.
     */
-    private _enter(name);
+    private _enter;
     /**
        * Sends all data needed for View.
        */
-    private _sendChanges();
+    private _sendChanges;
     /**
      * Handler for transitioend event
      */
@@ -170,7 +170,7 @@ export declare class StageComponent implements OnInit, OnDestroy {
     /**
        * Enters into a 'dragging' state
        */
-    private _enterDragging();
+    private _enterDragging;
     /**
      * Handles the end of 'animationend' event
      * @param id Id of slides
