@@ -15,6 +15,10 @@ export declare class AutoplayService implements OnDestroy {
      * Indicates whenever the autoplay is paused.
      */
     private _paused;
+    /**
+     * Shows whether the code (the plugin) changed the option 'AutoplayTimeout' for own needs
+     */
+    private _isArtificialAutoplayTimeout;
     private winRef;
     private docRef;
     constructor(carouselService: CarouselService, winRef: any, docRef: any);
@@ -35,11 +39,11 @@ export declare class AutoplayService implements OnDestroy {
        * @param speed - The animation speed for the animations.
        * @return
        */
-    private _getNextTimeout(timeout?, speed?);
+    private _getNextTimeout;
     /**
        * Sets autoplay in motion.
        */
-    private _setAutoPlayInterval();
+    private _setAutoPlayInterval;
     /**
      * Stops the autoplay.
      */
@@ -52,7 +56,11 @@ export declare class AutoplayService implements OnDestroy {
      * Manages by autoplaying according to data passed by _changedSettingsCarousel$ Obsarvable
      * @param data object with current position of carousel and type of change
      */
-    private _handleChangeObservable(data);
+    private _handleChangeObservable;
+    /**
+     * Starts autoplaying of the carousel in the case when user leaves the carousel before it starts translateing (moving)
+     */
+    private _playAfterTranslated;
     /**
      * Starts pausing
      */
